% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{cal.Activity}
\alias{cal.Activity}
\title{calculate activity for network targets}
\usage{
cal.Activity(all_target = NULL, cal_mat = NULL, es.method = "mean")
}
\arguments{
\item{es.method:}{mean, weighted mean, maxmean, absmean}
}
\description{
calculate activity for network targets
}
