% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{check_scalefree}
\alias{check_scalefree}
\title{functions to check the scale free feature of the network}
\usage{
check_scalefree(gr1)
}
\description{
functions to check the scale free feature of the network
}
