% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{combinePvalVector}
\alias{combinePvalVector}
\title{combine Pvalues by Fisher or Stouffer's method}
\usage{
combinePvalVector(pvals, method = c("Stouffer", "Fisher"),
  signed = TRUE, twosided = TRUE)
}
\arguments{
\item{twosided}{For Fisher's method: if twosided, pvalues are transformed into single}

\item{dat, }{a dataframe containing pvalue.cols}

\item{sign.cols}{column labels indicating sign of stat, if NULL, use absolute values to calculate overall stat for Stouffer's method}

\item{combine}{one pvalues vector, the signed version}
}
\description{
combine Pvalues by Fisher or Stouffer's method
}
