% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{combRowEvid.2grps}
\alias{combRowEvid.2grps}
\title{combRowEvid.2grps is a inner function for getDE.BID.2G.}
\usage{
combRowEvid.2grps(d, comp, family = gaussian, method = c("MLE",
  "Bayesian"), pooling = c("full", "no", "partial"), n.iter = 1000,
  prior.V.scale = 0.02, prior.R.nu = 1, prior.G.nu = 2,
  nitt = 13000, burnin = 3000, thin = 10, restand = TRUE,
  logTransformed = TRUE, log.base = 2,
  average.method = c("geometric", "arithmetic"), pseudoCount = 0)
}
\arguments{
\item{d}{input expression matrix}

\item{comp}{an indicator vector}

\item{family}{family Objects for Models, default is 'gaussian'}

\item{method}{choose from 'MLE' or 'Bayesian'}

\item{pooling}{choose from 'full','no','partial'}

\item{n.iter}{number of interaction, default is 1000}

\item{prior.V.scale}{default is 0.02}

\item{prior.R.nu}{default is 1}

\item{prior.G.nu}{default is 2}

\item{nitt}{default is 13000}

\item{burnin}{default is 3000}

\item{thin}{default is 10}

\item{restand}{default is TRUE}

\item{logTransformed}{default is TRUE}

\item{log.base}{default is 2}

\item{average.method}{choose from c('geometric','arithmetic')}

\item{pseudoCount}{default is 0}
}
\description{
combRowEvid.2grps is a inner function for getDE.BID.2G.
}
