% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{db.preload}
\alias{db.preload}
\title{Load database files used for NetBID2 into R workspace.}
\usage{
db.preload(use_level = "transcript", use_spe = "human",
  update = FALSE, TF_list = NULL, SIG_list = NULL,
  input_attr_type = "external_gene_name", main.dir = NULL,
  db.dir = sprintf("\%s/db/", main.dir))
}
\arguments{
\item{use_level}{character, either 'transcript' or 'gene', default is 'gene'}

\item{use_spe}{character, input the species name (e.g 'human', 'mouse', 'rat'), default is 'human'}

\item{update}{logical,whether to update if previous RData has been generated, default FALSE}

\item{TF_list}{a character vector,input the list of TF names, if NULL, will use pre-defined list in the package, default NULL}

\item{SIG_list}{a character vector,input the list of SIG names, if NULL, will use pre-defined list in the package, default NULL}

\item{input_attr_type}{character, input the type for the list of TF_list, SIG_list. If no input for TF_list, SIG_list, just leave it to NULL, default NULL.
See biomaRt \url{https://bioconductor.org/packages/release/bioc/vignettes/biomaRt/inst/doc/biomaRt.html} for more details.}

\item{main.dir}{character, main file path for NetBID2, if NULL, will set to \code{system.file(package = "NetBID2")}. Default is NULL.}

\item{db.dir}{character, file path for saving the RData, default is \code{db} directory under \code{main.dir} when setting for \code{main.dir}.}
}
\value{
Reture TRUE if success and FALSE if not. Will load two variables into R workspace, tf_sigs and db_info.
}
\description{
\code{db.preload} returns the TF (transcription factors) and Sig (signaling factors) list (tf_sigs)
and biomart database information (db_info) by input interested species name and make choice from
gene or transcript level.
}
\details{
This is a pre-processing function for NetBID2, user could input the species name (e.g human, mouse),
analysis level (transcript or gene level) and optionally input TF list or SIG list
(otherwise will use list from package data). The function could automatically download information
from biomart and save into RData under the db/ directory with specified species name and analysis level.
}
\examples{
db.preload(use_level='gene',use_spe='human',update=FALSE)

\dontrun{
db.preload(use_level='transcript',use_spe='human',update=FALSE)
db.preload(use_level='gene',use_spe='mouse',update=FALSE)
}
}
