% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{draw.2D}
\alias{draw.2D}
\title{Draw 2D dimension plot}
\usage{
draw.2D(X, Y, class_label, xlab = "PC1", ylab = "PC2",
  label_cex = 0.8, main = "", point_cex = 1)
}
\description{
Draw 2D dimension plot
}
