% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{draw.3D}
\alias{draw.3D}
\title{Visualize Sample Clustering Result in 3D Plot}
\usage{
draw.3D(
  X,
  Y,
  Z,
  class_label,
  xlab = "PC1",
  ylab = "PC2",
  zlab = "PC3",
  legend_cex = 0.8,
  main = "",
  point_cex = 1,
  legend_pos = "topright",
  legend_ncol = 1,
  use_color = NULL,
  pre_define = NULL,
  ...
)
}
\arguments{
\item{X}{a vector of numerics, the x coordinates of points in the plot. If user would like to create a PCA biplot, this parameter should be the first component.}

\item{Y}{a vector of numerics, the y coordinates of points in the plot. If user would like to create a PCA biplot, this parameter should be the second component.}

\item{Z}{a vector of numerics, the z coordinates of points in the plot. If user would like to create a PCA biplot, this parameter should be the third component.}

\item{class_label}{a vector of characters, labels or categories of samples. The vector name should be sample names.}

\item{xlab}{character, the label for x-axis. Default is "PC1".}

\item{ylab}{character, the label for y-axis. Default is "PC2".}

\item{zlab}{character, the label for z-axis. Default is "PC3".}

\item{legend_cex}{numeric, giving the amount by which the text of legend should be magnified relative to the default. Default is 0.8.}

\item{main}{character, an overall title for the plot. Default is "".}

\item{point_cex}{numeric, giving the amount by which the size of the data points should be magnified relative to the default. Default is 1.}

\item{legend_pos}{character, the position of legend. Default is "topright".}

\item{legend_ncol}{integer, number of columns of legend. Default is 1.}

\item{use_color}{a vector of color codes, colors to be assigned to each member of display label. Default is brewer.pal(9, 'Set1').}

\item{pre_define}{a vector of characters, pre-defined color codes for a certain input (e.g. c("blue", "red") with names c("A", "B")). Default is NULL.}

\item{...}{other paramters used in \code{scatter3D}.}
}
\value{
Return a logical value. If TRUE, the plot has been created successfully.
}
\description{
\code{draw.3D} creates a 3D plot to visualize the sample clustering result.
}
\examples{
mat1 <- matrix(rnorm(2000,mean=0,sd=1),nrow=100,ncol=20)
rownames(mat1) <- paste0('Gene',1:nrow(mat1))
colnames(mat1) <- paste0('Sample',1:ncol(mat1))
pc <- stats::prcomp(t(mat1))$x
pred_label <- kmeans(pc,centers=4)$cluster ## this can use other cluster results
draw.3D(X=pc[,1],Y=pc[,2],Z=pc[,3],class_label=pred_label)
}
