% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{draw.clustComp}
\alias{draw.clustComp}
\title{Draw the cluster comparison between predicted label and observed label.}
\usage{
draw.clustComp(pred_label, obs_label, strategy = "ARI", use_col = TRUE,
  low_K = 5, highlight_clust = NULL, main = NULL, clust_cex = 1,
  outlier_cex = 0.3)
}
\arguments{
\item{pred_label}{a vector of characters, the predicted label.}

\item{obs_label}{a vector of characters, the observed label}

\item{strategy}{character, the strategy to compare with labels,
choose from 'ARI (adjusted rand index)', 'NMI (normalized mutual information)', 'Jaccard'. Default is 'ARI'.}

\item{use_col}{logical, whether or not to use color in the plot. Default is TRUE.}

\item{low_K}{integer, the lowest number to display on the figures. Number smaller than this will directly display the sample name.
Default is 5}

\item{highlight_clust}{a vector of characters, the cluster need to be highlighted on the plot.}

\item{main}{character, the title for the plot.}

\item{clust_cex}{numeric, cex for the cluster label. Default is 1}

\item{outlier_cex}{numeric, cex for the sample names. Default is 0.3}
}
\value{
a matrix for the number in the plot
}
\description{
\code{draw.clustComp} is a function to draw the comparison between the predicted label and observed label.
}
\examples{
network.par <- list()
network.par$out.dir.DATA <- system.file('demo1','network/DATA/',package = "NetBID2")
NetBID.loadRData(network.par=network.par,step='exp-QC')
mat <- exprs(network.par$net.eset)
phe <- pData(network.par$net.eset)
intgroup <- 'subgroup'
pred_label <- draw.pca.kmeans(mat=mat,all_k = NULL,
                             obs_label=get_obs_label(phe,intgroup),
                             kmeans_strategy='consensus')
draw.clustComp(pred_label,get_obs_label(phe,intgroup),outlier_cex=1,low_K=2,use_col=TRUE)
draw.clustComp(pred_label,get_obs_label(phe,intgroup),outlier_cex=1,low_K=2,use_col=FALSE)
}
