% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{draw.eset.QC}
\alias{draw.eset.QC}
\title{QC plot for ExpressionSet class object.}
\usage{
draw.eset.QC(eset, outdir = ".", do.logtransform = FALSE,
  intgroup = NULL, prefix = "", choose_plot = c("heatmap", "pca",
  "density", "meansd"))
}
\arguments{
\item{eset}{ExpressionSet class, the input ExpressionSet class object to be plot.}

\item{outdir}{character, output directory to save the figures.
Suggest to set \code{network.par$out.dir.QC} or \code{analysis.par$out.dir.QC}}

\item{do.logtransform}{logical, whether to do log transformation before drawing the QC plots. Default is FALSE.}

\item{intgroup}{a vector of characters, the interested groups from the phenotype information of the eset to be used in plot.
If NULL, will automatcially extract all possible groups by \code{get_int_group}.
Default is NULL.}

\item{prefix}{character, the prefix for the QC figure name.Default is "".}

\item{choose_plot}{a vector of characters,
choose one or multiple from 'heatmap', 'pca', 'density', 'meansd.'
Default is 'heatmap','pca','density','meansd'.}
}
\description{
\code{draw.eset.QC} is a function to draw the basic QC plots for an ExpressionSet class object.
The QC plots include heatmap, pca, density and meansd.
}
\examples{
\dontrun{
network.par <- list()
network.par$out.dir.DATA <- system.file('demo1','network/DATA/',package = "NetBID2")
NetBID.loadRData(network.par=network.par,step='exp-QC')
intgroups <- get_int_group(network.par$net.eset)
network.par$out.dir.QC <- getwd() ## set the output directory
draw.eset.QC(network.par$net.eset,outdir=network.par$out.dir.QC,intgroup=intgroups)
}
}
