% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{draw.GSEA.NetBID.GS}
\alias{draw.GSEA.NetBID.GS}
\title{GSEA (gene set enrichment analysis) plot for the list of gene sets.}
\usage{
draw.GSEA.NetBID.GS(DE = NULL, name_col = NULL, profile_col = NULL,
  profile_trend = "pos2neg", sig_gs_list = NULL, gs_DA_Z = NULL,
  use_gs2gene = NULL, top_gs_number = 30, target_col = "RdBu",
  left_annotation = "", right_annotation = "", main = "",
  profile_sig_thre = 0, Z_sig_thre = 1.64, pdf_file = NULL)
}
\arguments{
\item{DE}{data.frame,the differentiated expression results.
This data.frame could be generated by using \code{getDE.limma.2G} or \code{getDE.BID.2G}.
If user want to generate this data.frame by other strategies, the rownames must be the gene names and
must contain the columns indicating the differentiated expression profile.}

\item{name_col}{character, the name of the column in \code{DE}, which contains the gene name. If NULL, will use the rownames of \code{DE}.
Default is NULL.}

\item{profile_col}{character, the name of the column in \code{DE}, which will be used as the differentiated expression profile.
If DE is created by \code{getDE.limma.2G} or \code{getDE.BID.2G}, this parameter could be 'logFC' or 't'.}

\item{profile_trend}{character, the choice of how to display the profile, from high/positive to low/negative ('pos2neg')
or low/negative to high/positive ('neg2pos').Default is 'pos2neg'.}

\item{sig_gs_list}{a vector of characters, the name for the top gene sets.}

\item{gs_DA_Z}{a vector of numeric values, the Z statistics of differentiated activity (DA) for the sig_gs_list.
Better to give name to the vector, otherwise will automatically use sig_gs_list as the name.}

\item{use_gs2gene}{a list for geneset to genes, the name for the list is the gene set name and the content in each list is the vector for genes belong to that gene set.
This parameter could be obtained by choosing one from \code{all_gs2gene[['CP:KEGG']]}, or merge several categories by \code{merge_gs}.}

\item{top_gs_number}{numeric, number for the top significant gene sets to be displayed on the plot. Default is 30.}

\item{target_col}{character, choice of color pattern used to display the targets. Two options,'black' or 'RdBu'.
If set to 'black', the lines will be colored in black.
If set to 'RdBu', the lines will be colored into Red to Blue color bar. The color for the annotated genes is set according
to its value in the differentiated expression profile, with significant high set for red and low for blue.
The significant threshold is set by \code{profile_sig_thre}.
Default is 'RdBu'.}

\item{left_annotation}{character, annotation displayed on the left of the figure representing left condition of the rank_profile. Default is "".}

\item{right_annotation}{character, annotation displayed on the right of the figure representing right condition of the rank_profile. Default is "".}

\item{main}{character, title for the plot. Default is "".}

\item{profile_sig_thre}{numeric, threshold for the absolute values in profile to be treated as significance.
annotated genes without signifcant values in the profile will be colored in grey. Only work when \code{target_col_type} is set as "DE" and \code{target_col} is set as "RdBu".
Default is 0.}

\item{Z_sig_thre}{numeric, threshold for the Z statistics in \code{driver_DA_Z} and \code{driver_DE_Z} to be treated as signifcance.
Only signifcant values will have background color. Default is 1.64.}

\item{pdf_file}{character, file path for the pdf file to save the figure into pdf format.If NULL, will not generate pdf file. Default is NULL.}
}
\value{
logical value indicating whether the plot has been successfully generated
}
\description{
\code{draw.GSEA.NetBID.GS} will generate a GSEA plot for the list of gene sets, including the annotated genes' position on the differentiated expression profile, with
statistics of differentiated activity (DA) for each gene set.
}
\details{
This is a plot function to draw GSEA for the list of gene sets by the input of differentiated expression profile.
User could choose to display the annotated genes by selecting black color or red to blue color bar.
}
\examples{
\dontrun{
db.preload(use_level='transcript',use_spe='human',update=FALSE)

## get all_gs2gene

analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo','gene set/DATA/',package = "NetBID2")
NetBID2.loadRData(analysis.par=analysis.par,step='ms-tab')

ms_tab <- analysis.par$final_ms_tab
comp <- 'G4.Vs.others'
DE <- analysis.par$DE[[comp]]
analysis.par$out.dir.PLOT <- getwd()

## directory for saving the pdf files
exp_mat_gene <- exprs(analysis.par$cal.eset)

## calculate activity for all genesets
use_gs2gene <- merge_gs(all_gs2gene=all_gs2gene,
                       use_gs=c('H','CP:BIOCARTA','CP:REACTOME','CP:KEGG','C5'))
ac_gs <- cal.Activity.GS(use_gs2gene = use_gs2gene,cal_mat = exp_mat_gene)

## get DA for the gene set
phe_info <- pData(analysis.par$cal.eset)
G0  <- rownames(phe_info)[which(phe_info$`subgroup`!='G4')]
# get sample list for G0
G1  <- rownames(phe_info)[which(phe_info$`subgroup`=='G4')]
# get sample list for G1
DA_gs <- getDE.limma.2G(eset=generate.eset(ac_gs),G1=G1,G0=G0,
                        G1_name='G4',G0_name='others')
## or use: DA_gs <- getDE.BID.2G(eset=generate.eset(ac_gs),G1=G1,G0=G0,
                        G1_name='G4',G0_name='others')
## draw vocalno plot for top sig-GS
sig_gs <- draw.volcanoPlot(dat=DA_gs,
                           label_col='ID',
                           logFC_col='logFC',
                           Pv_col='P.Value',
                           logFC_thre=0.25,
                           Pv_thre=1e-4,
                           main='Volcano Plot for gene sets',
                           show_label=TRUE,
                           label_type = 'distribute',
                           label_cex = 0.5,
                           pdf_file=sprintf('\%s/vocalno_GS_DA.pdf',
                           analysis.par$out.dir.PLOT))
## GSEA plot for the significant gene sets
draw.GSEA.NetBID.GS(DE=DE,name_col='ID',
                    profile_col='t',profile_trend='pos2neg',
                    sig_gs_list = sig_gs$names,
                    gs_DA_Z=DA_gs[sig_gs$names,'Z-statistics'],
                    use_gs2gene = use_gs2gene,
                    top_gs_number=5,target_col='RdBu',
                    left_annotation = 'test_left',
                    right_annotation = 'test_right',
                    main='test',Z_sig_thre=1.64,profile_sig_thre = 0,
                    pdf_file=sprintf('\%s/NetBID_GSEA_GS_demo1.pdf',
                    analysis.par$out.dir.PLOT))
}
}
