% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{draw.NetBID}
\alias{draw.NetBID}
\title{Create Plot to show Differential Expression and Differential Activity Analysis of Top Drivers}
\usage{
draw.NetBID(DA_list = NULL, DE_list = NULL, main_id = NULL,
  top_number = 30, DA_display_col = "P.Value",
  DE_display_col = "logFC", z_col = "Z-statistics", digit_num = 2,
  row_cex = 1, column_cex = 1, text_cex = 1, col_srt = 60,
  pdf_file = NULL)
}
\arguments{
\item{DA_list}{list, contains the differential activity (DA) analysis result.}

\item{DE_list}{list, contains the differential expression (DE) analysis result.}

\item{main_id}{character, the top genes/drivers from which DA comparison group. Must be one of the names in \code{DA_list}.
If NULL, the first element name of \code{DA_list} will be used. Default is NULL.}

\item{top_number}{integer, the number of the top significant genes/drivers to be displayed in the plot.
Default is 30.}

\item{DA_display_col}{character, which statistic column from NetBID analysis is to be used as the column of the left heatmap.
Default is "P.Value".}

\item{DE_display_col}{character, which statistic column from NetBID analysis is to be used as the column of the right heatmap.
Default is "logFC".}

\item{z_col}{character, which statistic column from NetBID analysis is to be used as the color scale of the heatmap.
Default is "Z-statistics".}

\item{digit_num}{integer, indicating the number of decimal places (round) or significant digits (signif) to be used.
Default is 2.}

\item{row_cex}{numeric, giving the amount by which the text of row names should be magnified relative to the default.
Default is 1.}

\item{column_cex}{numeric, giving the amount by which the text of column names should be maginified relative to the default.
Default is 1.}

\item{text_cex}{numeric, giving the amount by which the text of in the table cell should be maginified relative to the default.
Default is 1.}

\item{col_srt}{numeric, rotation angle of the column labels at the bottom of heatmap.
Default is 60.}

\item{pdf_file}{character, the path to save the plot as PDF file.
If NULL, PDF file will not be generated. Default is NULL.}
}
\value{
Return a logical value. If TRUE, the plot is created successfully.
}
\description{
\code{draw.NetBID} creates two side-by-side heatmaps to show the result of NetBID analysis.
Both the differential expression analysis (the right heatmap) and differential activity analysis (the left heatmap) of top drivers are shown in the plot.
}
\examples{
analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo1','driver/DATA/',package = "NetBID2")
NetBID.loadRData(analysis.par=analysis.par,step='ms-tab')
draw.NetBID(DA_list=analysis.par$DA,DE_list=analysis.par$DE)
}
