% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{draw.network.QC}
\alias{draw.network.QC}
\title{Generate QC plot for the network object}
\usage{
draw.network.QC(igraph_obj, outdir = NULL, prefix = "",
  directed = TRUE, weighted = FALSE, generate_html = TRUE)
}
\arguments{
\item{igraph_obj}{igraph object, this could be generated by \code{get.SJAracne.network}}

\item{outdir}{character, the output directory to save the QC figures.}

\item{prefix}{character, the prefix for the QC figure name.Default is "".}

\item{directed}{logical, whether to treat the network as directed network. Default is TRUE.}

\item{weighted}{logical, whether to treat the network as weighted network. Default is FALSE.}

\item{generate_html}{logical, whether to generate html file for the report.
If TRUE, will generate a html file by rmarkdown, otherwise will generate separate pdf files.
Default is TRUE.}
}
\value{
logical value indicating whether the plot has been successfully generated
}
\description{
\code{draw.network.QC} is a function to draw the QC plot for the network object,
mainly the density plot for the degree and the check for scale-free feature.
}
\examples{
\dontrun{
if(exists('analysis.par')==TRUE) rm(analysis.par)
network.dir <- sprintf('\%s/demo1/network/',system.file(package = "NetBID2")) # use demo
network.project.name <- 'project_2019-02-14' # demo project name
project_main_dir <- 'test/'
project_name <- 'test_driver'
analysis.par  <- NetBID.analysis.dir.create(project_main_dir=project_main_dir,
                                            prject_name=project_name,
                                            network_dir=network.dir,
                                            network_project_name=network.project.name)
analysis.par$tf.network  <- get.SJAracne.network(network_file=analysis.par$tf.network.file)
analysis.par$out.dir.PLOT <- getwd() ## directory for saving the pdf files
draw.network.QC(analysis.par$tf.network$igraph_obj,
                outdir=analysis.par$out.dir.QC,prefix='TF_net_')
}
}
