% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{draw.oncoprint}
\alias{draw.oncoprint}
\title{Draw Oncoprint Plot to display the mutation information}
\usage{
draw.oncoprint(
  phenotype_info = NULL,
  Missense_column = NULL,
  Missense_label = NULL,
  Amplification_column = NULL,
  Amplification_label = NULL,
  Deletion_column = NULL,
  Deletion_label = NULL,
  Sample_column = NULL,
  main = "",
  pdf_file = NULL,
  ...
)
}
\arguments{
\item{phenotype_info}{data.frame, phenotype of samples. Users can call \code{Biobase::pData(eset)} to create.}

\item{Missense_column}{character, column names from \code{phenotype_info}.}

\item{Missense_label, }{character, gene label for the Missense_column.}

\item{Amplification_column}{character, column names from \code{phenotype_info}.}

\item{Amplification_label, }{character, gene label for the Amplification_column.}

\item{Deletion_column}{character, column names from \code{phenotype_info}.}

\item{Deletion_label, }{character, gene label for the mDeletion_column.}

\item{Sample_column}{character, column names from \code{phenotype_info}. If not NULL, will show sample names in the figure.}

\item{main}{character, an overall title for the plot. Default is "".}

\item{pdf_file}{character, the file path to save plot as PDF file. If NULL, no PDF file will be saved. Default is NULL.}

\item{..., }{for more options, please check \code{?oncoPrint} for more details.}
}
\value{
Return a logical value. If TRUE, the plot has been created successfully.
}
\description{
\code{draw.oncoprint} plots the heatmap to display the mutation information for samples.
}
\examples{
all_sample <- sprintf('Sample\%s',1:30)
group1_sample <- sample(all_sample,18) ## demo sample for KRAS missense mutation
group2_sample <- sample(all_sample,12) ## demo sample for MYC amplification
group3_sample <- sample(all_sample,10) ## demo sample for MYC missense mutation
group4_sample <- sample(all_sample,1) ## demo sample for MYC deletion
phenotype_info_demo <-
  data.frame(sample =sprintf('Sample\%s',1:30),
             KRAS_MIS=ifelse(all_sample \%in\% group1_sample,1,0),
             MYC_AMP=ifelse(all_sample \%in\% group2_sample,1,0),
             MYC_MIS=ifelse(all_sample \%in\% group3_sample,1,0),
             MYC_DEL=ifelse(all_sample \%in\% group4_sample,1,0))
draw.oncoprint(phenotype_info=phenotype_info_demo,
               Missense_column=c('KRAS_MIS','MYC_MIS'),Missense_label=c('KRAS','MYC'),
               Amplification_column=c('MYC_AMP'),Amplification_label=c('MYC'),
               Deletion_column=c('MYC_DEL'),Deletion_label=c('MYC'),
               Sample_column='sample',
               main="OncoPrint for the demo dataset")
\dontrun{
}
}
