% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{draw.volcanoPlot}
\alias{draw.volcanoPlot}
\title{Draw Volcano Plot for Top DE (differentiated expressed) Genes or DA (differentiated activity) Drivers}
\usage{
draw.volcanoPlot(dat = NULL, label_col = NULL, logFC_col = NULL,
  Pv_col = NULL, logFC_thre = 1.5, Pv_thre = 0.01,
  show_plot = TRUE, xlab = "log2 Fold Change", ylab = "P-value",
  show_label = FALSE, label_cex = 0.5, legend_cex = 0.8,
  label_type = "distribute", main = "", pdf_file = NULL)
}
\arguments{
\item{dat}{data.frame, the master table created by function \code{generate.masterTable}. Or a table with columns of the following parameters.}

\item{label_col}{character, the name of the column in \code{dat} contains gene/driver names.}

\item{logFC_col}{character, the name of the column in \code{dat} contains logFC values.}

\item{Pv_col}{character, the name of the column in \code{dat} contains P-values.}

\item{logFC_thre}{numeric, the threshold of logFC. Genes or drivers with absolute logFC value higher than the threshold will be kept.
Default is 1.5.}

\item{Pv_thre}{numeric, the threshold of P-values. Genes or drivers with P-values lower than the threshold will be kept.
Default is 0.01.}

\item{show_plot}{logical, if TRUE, the plot will be shown in the plot pane. Default is TRUE.}

\item{xlab}{character, a title for the X axis.}

\item{ylab}{character, a title for the Y axis.}

\item{show_label}{logical, if TRUE, labels of selected genes or drivers will be displayed in the plot. Default is FALSE.}

\item{label_cex}{numeric, giving the amount by which the text of genes/drivers label should be magnified relative to the default. Default is 0.5.}

\item{legend_cex}{numeric, giving the amount by which the text of legend should be magnified relative to the default. Default is 0.7.}

\item{label_type}{character, users can choose between "origin" and "distribute". If "origin", all the labels will be displayed without location modification.
If "distribute", location of labels will be rearranged to avoid overlap. Default is "distribute".}

\item{main}{character, an overall title for the plot.}

\item{pdf_file}{character, the path to save the plot as PDF file. If NULL, no PDF file will be created. Default is NULL.}
}
\value{
Return a data.frame of selected significant genes or drivers, with columns contain \code{label_col}, \code{logFC_col} and \code{Pv_col}.
}
\description{
\code{draw.volcanoPlot} draws the volcano plot to quickly identify DE genes or DA drivers with large fold change and significant P-value in large data sets.
And return a data.frame of these highlighted genes/drivers.
}
\details{
Top genes or drivers will be colored (blue for down-regulated and red for up-regulated) and labeled with their names.
This function requires the input of master table and two thresholds of logFC and P-value.
}
\examples{
analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo1','driver/DATA/',package = "NetBID2")
NetBID.loadRData(analysis.par=analysis.par,step='ms-tab')
ms_tab <- analysis.par$final_ms_tab
sig_driver <- draw.volcanoPlot(dat=ms_tab,label_col='gene_label',
                               logFC_col='logFC.G4.Vs.others_DA',
                               Pv_col='P.Value.G4.Vs.others_DA',
                               logFC_thre=0.4,
                               Pv_thre=1e-7,
                               main='Volcano Plot for G4.Vs.others_DA',
                               show_label=FALSE,
                               label_type = 'origin',
                               label_cex = 0.5)

\dontrun{
analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo1','driver/DATA/',package = "NetBID2")
NetBID.loadRData(analysis.par=analysis.par,step='ms-tab')
ms_tab <- analysis.par$final_ms_tab
analysis.par$out.dir.PLOT <- getwd() ## directory for saving the pdf files
sig_driver <- draw.volcanoPlot(dat=ms_tab,label_col='gene_label',
                               logFC_col='logFC.G4.Vs.others_DA',
                               Pv_col='P.Value.G4.Vs.others_DA',
                               logFC_thre=0.4,
                               Pv_thre=1e-7,
                               main='Volcano Plot for G4.Vs.others_DA',
                               show_label=FALSE,
                               label_type = 'origin',
                               label_cex = 0.5)
sig_gene <- draw.volcanoPlot(dat=ms_tab,label_col='geneSymbol',
                               logFC_col='logFC.G4.Vs.others_DE',
                               Pv_col='P.Value.G4.Vs.others_DE',
                               logFC_thre=2,Pv_thre=1e-3,
                               main='Volcano Plot for G4.Vs.others_DE',
                               show_label=FALSE,
                               pdf_file=sprintf('\%s/vocalno_nolabel_DE.pdf',
                               analysis.par$out.dir.PLOT))
}
}
