% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{find.gsByGene}
\alias{find.gsByGene}
\title{find.gsByGene}
\usage{
find.gsByGene(gene = NULL, use_gs = NULL)
}
\arguments{
\item{gene}{character}

\item{use_gs}{a vector of characters}
}
\description{
find.gsByGene
}
