% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{funcEnrich.GSEA}
\alias{funcEnrich.GSEA}
\title{Gene Set Enrichment Analysis by GSEA}
\usage{
funcEnrich.GSEA(
  rank_profile = NULL,
  use_gs = NULL,
  gs2gene = NULL,
  min_gs_size = 5,
  max_gs_size = 500,
  Pv_adj = "fdr",
  Pv_thre = 0.1,
  test_strategy = "GSEA",
  nperm = 1000,
  use_seed = 999
)
}
\arguments{
\item{rank_profile}{a named vector of numerics, the differential values (DE or DA) calculated from a sample comparison (e.g. "G4 vs. Others").
Names of the vector must be gene names.
For the DA, user could use 'processDriverProfile()' to convert the DA profile into gene-name based profile.
The differential values can be "logFC" or "t-statistics".}

\item{use_gs}{a vector of characters, the names of gene sets.
If \code{gs2gene} is NULL, \code{all_gs2gene} will be used. The \code{use_gs} must be the subset of \code{names(all_gs2gene)}.
If "all", all the gene sets in \code{gs2gene} will be used.
If user input his own \code{gs2gene} list, \code{use_gs} will be set to "all" as default.
Default is c("H", "CP:BIOCARTA", "CP:REACTOME", "CP:KEGG").}

\item{gs2gene}{list, a list contains elements of gene sets.
The name of the element is gene set, each element contains a vector of genes in that gene set.
If NULL, will use \code{all_gs2gene}, which is created by function \code{gs.preload}. Default is NULL.}

\item{min_gs_size}{numeric, the minimum size of gene set to analysis. Default is 5.}

\item{max_gs_size}{numeric, the maximum size of gene set to analysis. Default is 500.}

\item{Pv_adj}{character, method to adjust P-value. Default is "fdr".
For details, please check \code{p.adjust.methods}.}

\item{Pv_thre}{numeric, threshold for the adjusted P-values. Default is 0.1.}

\item{test_strategy}{choose from "KS" and "GSEA". Default is "GSEA".
If "KS", will perform a Kolmogorov-Smirnov test to get the significance value.}

\item{nperm}{numeric, number of random permutations. Default is 1000.
This function only do gene-label based permutation reshuffling.}

\item{use_seed}{integer, the random seed. Default is 999.}
}
\value{
Return a data.frame, contains gene sets with significant enrichment statistics.
Column details are as follows (test_strategy=GSEA),

\item{#Name}{Name of the enriched gene set}
\item{Total_item}{Size in the profile}
\item{Num_item}{Number of genes in the gene set (filtered by the profile list)}
\item{Ori_P}{Original P-value from GSEA Test}
\item{Adj_P}{Adjusted P-value}
\item{ES}{Enrichment Score}
\item{NES}{normalized Enrichment Score}
}
\description{
\code{funcEnrich.GSEA} performs gene set enrichment analysis to the input gene list, by using the GSEA Test.
}
\examples{
analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo1','driver/DATA/',package = "NetBID2")
NetBID.loadRData(analysis.par=analysis.par,step='ms-tab')
ms_tab <- analysis.par$final_ms_tab
## get significant gene set by driver's DA profile
DA_profile <- processDriverProfile(Driver_name=ms_tab$gene_label,
                                    Driver_profile=ms_tab$logFC.G4.Vs.others_DA,
                                    choose_strategy='absmax',
                                    return_type ='gene_statistics')
res1 <- funcEnrich.GSEA(rank_profile=DA_profile,
                         use_gs=c('H'),
                         Pv_thre=0.1,Pv_adj = 'none')
\dontrun{
}
}
