% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{generate.masterTable}
\alias{generate.masterTable}
\title{Generate master}
\usage{
generate.masterTable(use_comp = NULL, DE = NULL, DA = NULL,
  network = NULL, main_id_type = NULL, tf_sigs = tf_sigs,
  z_col = NULL, display_col = NULL)
}
\description{
Generate master
}
