% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{generate.masterTable.TF_SIG}
\alias{generate.masterTable.TF_SIG}
\title{Generate master table by input of TF and SIG results.}
\usage{
generate.masterTable.TF_SIG(use_comp = NULL, DE = NULL, DA_TF = NULL,
  DA_SIG = NULL, TF_network = NULL, SIG_network = NULL,
  main_id_type = NULL, tf_sigs = tf_sigs, z_col = NULL,
  display_col = NULL)
}
\description{
Generate master table by input of TF and SIG results.
}
