% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{get.class.color}
\alias{get.class.color}
\title{Generate a color vector for input character.}
\usage{
get.class.color(x, use_color = NULL, pre_define = NULL,
  use_inner = TRUE)
}
\arguments{
\item{x}{a vector of characters.}

\item{use_color}{a vector of characters, color bar used to generate the color vector for the input.Default is brewer.pal(12, 'Set3').}

\item{pre_define}{a vector of characters, pre-defined color code for some of the input characters. Default is NULL.}

\item{user_inner}{logical, indicating whether to use inner pre-defined color code for some characters. Default is TRUE.}
}
\value{
Will return a vector of colors with names the input vector characters.
}
\description{
\code{get.class.color} will generate a vector of colors for input character.
}
\details{
This is a simple function to generate a vector of colors for input characters.
Users could define some of the colors for part of the inputs.
}
\examples{
get.class.color(c('ClassA','ClassB','ClassC','ClassA','ClassC','ClassC'))
get.class.color(c('ClassA','ClassB','ClassC','SHH','WNT','Group3','Group4'),
                use_inner=FALSE)
get.class.color(c('ClassA','ClassB','ClassC','SHH','WNT','Group3','Group4'),
                use_inner=FALSE,use_color=brewer.pal(8, 'Set1'))

pre_define <- c('blue', 'red', 'yellow', 'green','yellow', 'green')
                ## pre-defined colors for MB
names(pre_define) <- c('WNT', 'SHH', 'Group3', 'Group4','GroupC', 'GroupD')
                ##pre-defined color name for MB
get.class.color(c('ClassA','ClassB','ClassC','SHH','WNT','Group3','Group4'),
                pre_define=pre_define,use_inner=FALSE)

\dontrun{
}
}
