% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{get_clustComp}
\alias{get_clustComp}
\title{Get Score between predicted label and observed label.}
\usage{
get_clustComp(pred_label, obs_label, strategy = "ARI")
}
\arguments{
\item{pred_label}{a vector of characters, the predicted label.}

\item{obs_label}{a vector of characters, the observed label}

\item{strategy}{character, the strategy to compare with labels,
choose from 'ARI (adjusted rand index)', 'NMI (normalized mutual information)', 'Jaccard'. Default is 'ARI'.}
}
\value{
score for the comparison
}
\description{
\code{get_clustComp} is a function to compare the predicted label and observed label and return the score.
}
\examples{
obs_label <- c('A','A','A','B','B','C','D')
pred_label  <- c(1,1,1,1,2,2,2)
get_clustComp(pred_label,obs_label)
}
