% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{get_IDtransfer}
\alias{get_IDtransfer}
\title{Gene ID conversion related functions.}
\usage{
get_IDtransfer(from_type = NULL, to_type = NULL, use_genes = NULL,
  dataset = db_info[1])
}
\arguments{
\item{from_type}{character, attribute name from the biomaRt package, such as 'ensembl_gene_id','ensembl_gene_id_version'
'ensembl_transcript_id','ensembl_transcript_id_version','refseq_mrna'. Full list could be obtained by
\code{listAttributes(mart)$name}, where \code{mart} is the output of \code{useMart} function.
The type must be the gene type for the input \code{use_genes}.}

\item{to_type}{character, character, attribute name from the biomaRt package, the gene type will be convereted to.}

\item{use_genes}{a vector of characters, gene list used for ID conversion.}

\item{dataset}{character, name for the dataset used for ID conversion, such as 'hsapiens_gene_ensembl'.
If NULL, will use \code{db_info[1]} if run \code{db.preload} brefore. Default is NULL.}
}
\value{
\code{get_IDtransfer} will return a data.frame for the transfer table.
}
\description{
\code{get_IDtransfer} will generate a transfer table for ID conversion by input the from-gene type and to-gene type.
}
\examples{
use_genes <- c("ENST00000210187","ENST00000216083","ENST00000216127",
             "ENST00000216416","ENST00000217233","ENST00000221418")
transfer_tab <- get_IDtransfer(from_type = 'ensembl_transcript_id',
                               to_type='external_gene_name',
                               use_genes=use_genes,
                               dataset='hsapiens_gene_ensembl')
                               ## get transfer table !!!
res1 <- get_name_transfertab(use_genes,transfer_tab=transfer_tab)
transfer_tab_withtype <- get_IDtransfer2symbol2type(from_type = 'ensembl_transcript_id',
                                                   use_genes=use_genes,
                                                   dataset='hsapiens_gene_ensembl')
                                                   ## get transfer table !!!
\dontrun{
}
}
