% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{get_IDtransfer_betweenSpecies}
\alias{get_IDtransfer_betweenSpecies}
\title{Gene ID conversion related functions.}
\usage{
get_IDtransfer_betweenSpecies(from_spe = "human", to_spe = "mouse",
  from_type = NULL, to_type = NULL, use_genes = NULL)
}
\arguments{
\item{from_spe}{character, input the species name (e.g 'human', 'mouse', 'rat') that the \code{use_genes} belong to, default is 'human'}

\item{to_spe}{character, input the species name (e.g 'human', 'mouse', 'rat') that need to transfered to, default is 'mouse'}

\item{from_type}{character, attribute name from the biomaRt package, such as 'ensembl_gene_id','ensembl_gene_id_version'
'ensembl_transcript_id','ensembl_transcript_id_version','refseq_mrna'. Full list could be obtained by
\code{listAttributes(mart)$name}, where \code{mart} is the output of \code{useMart} function.
The type must be the gene type for the input \code{use_genes}.}

\item{to_type}{character, character, attribute name from the biomaRt package, the gene type will be convereted to.}

\item{use_genes}{a vector of characters, gene list used for ID conversion. Must be the genes with \code{from_type} in \code{from_spe}.}
}
\value{
\code{get_IDtransfer_betweenSpecies} will return a data.frame for the transfer table.
}
\description{
\code{get_IDtransfer_betweenSpecies} will generate a transfer table for ID conversion between species.
}
\examples{
use_genes <- c("ENST00000210187","ENST00000216083","ENST00000216127",
             "ENST00000216416","ENST00000217233","ENST00000221418")
transfer_tab <- get_IDtransfer_betweenSpecies(from_spe='human',
                               to_spe='mouse',
                               from_type = 'ensembl_transcript_id',
                               to_type='external_gene_name',
                               use_genes=use_genes)
                               ## get transfer table !!!
transfer_tab <- get_IDtransfer_betweenSpecies(from_spe='human',
                               to_spe='mouse',
                               from_type = 'ensembl_transcript_id',
                               to_type='ensembl_transcript_id_version',
                               use_genes=use_genes)
                               ## get transfer table !!!
\dontrun{
}
}
