% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{get_int_group}
\alias{get_int_group}
\title{Get interested groups from the pData of an ExpressionSet object.}
\usage{
get_int_group(eset)
}
\arguments{
\item{eset, }{an ExpressionSet object, the input object for analysis.}
}
\value{
a vector of characters, the column names which could be used for sample cluster analysis
}
\description{
\code{get_int_group} is a simple function to extract columns with unique sample feature ranges from 2 to sample size-1.
}
\examples{
network.par <- list()
network.par$out.dir.DATA <- system.file('demo1','network/DATA/',package = "NetBID2")
NetBID.loadRData(network.par=network.par,step='exp-QC')
intgroups <- get_int_group(network.par$net.eset)
}
