% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{get_jac}
\alias{get_jac}
\title{get jaccard accuracy between predicted label and observed label}
\usage{
get_jac(pred_label, obs_label)
}
\description{
get jaccard accuracy between predicted label and observed label
}
