% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{get_net2target_list}
\alias{get_net2target_list}
\title{get target list by input network data}
\usage{
get_net2target_list(net_dat = NULL)
}
\description{
get target list by input network data
}
