% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{get_obs_label}
\alias{get_obs_label}
\title{Create a vector of each sample's selected phenotye descriptive information.}
\usage{
get_obs_label(phe_info, use_col, collapse = "|")
}
\arguments{
\item{phe_info}{data.frame, the phenotype data of the samples.
It is a data frame that can store any number of descriptive columns (covariates) for each sample row.
To get the phenotype data, using the accessor function \code{pData}.}

\item{use_col}{a vector of numerics or characters.
Users can select the interested descriptive column(s) by calling index or name of the column(s).}

\item{collapse}{character, an optional character string to separate the results when the length
of \code{use_col} is more than 1. Not NA_character. Default is "|".}
}
\value{
Return a vector of selected phenotype descriptive information (covariates) for each sample.
Vector name is the sample name.
}
\description{
\code{get_obs_label} creates a vector of each sample's selected phenotype descriptive information.
This is a helper function for data visualization.
}
\examples{
analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo1','driver/DATA/',package = "NetBID2")
NetBID.loadRData(analysis.par=analysis.par,step='ms-tab')
phe_info <- pData(analysis.par$cal.eset)
use_obs_class <- get_obs_label(phe_info = phe_info,'subgroup')
print(use_obs_class)
\dontrun{
}
}
