% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{getDE.BID.2G}
\alias{getDE.BID.2G}
\title{Bayesian Inference to get differential expression/activity between case and control}
\usage{
getDE.BID.2G(eset, G1 = NULL, G0 = NULL, G1_name = NULL,
  G0_name = NULL, verbose = TRUE)
}
\description{
Bayesian Inference to get differential expression/activity between case and control
}
