% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{getDE.limma.2G}
\alias{getDE.limma.2G}
\title{Limma functions to get differential expression/activity between case and control}
\usage{
getDE.limma.2G(eset = NULL, G1 = NULL, G0 = NULL, G1_name = NULL,
  G0_name = NULL, verbose = TRUE, random_effect = NULL)
}
\description{
Limma functions to get differential expression/activity between case and control
}
