% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{gs.preload}
\alias{gs.preload}
\title{Load MSigDB Database into R Workspace}
\usage{
gs.preload(
  use_spe = "Homo sapiens",
  update = FALSE,
  main.dir = NULL,
  db.dir = sprintf("\%s/db/", main.dir)
)
}
\arguments{
\item{use_spe}{character, name of interested species (e.g. "Homo sapiens", "Mus musculus").
Users can call \code{msigdbr_show_species()} to access the full list of available species names.
Default is "Homo sapiens".}

\item{update}{logical, if TRUE, the previous loaded RData will be updated. Default is FALSE.}

\item{main.dir}{character, the main file path of user's NetBID2 project.
If NULL, will be set to \code{system.file(package = "NetBID2")}. Default is NULL.}

\item{db.dir}{character, the file path to save the RData. Default is \code{db} directory under the \code{main.dir}, if one has a \code{main.dir}.}
}
\value{
Reture a logical value. If TRUE, MsigDB database is loaded successfully, with \code{all_gs2gene} and \code{all_gs2gene_info} created
in the workspace.
}
\description{
\code{gs.preload} downloads data from MSigDB and stores it into two variables in R workspace, \code{all_gs2gene} and \code{all_gs2gene_info}.
\code{all_gs2gene} is a list object with elements of gene sets collections.
\code{all_gs2gene_info} is a data.frame contains the description of each gene sets.
}
\details{
This is a pre-processing function for NetBID2 advanced analysis. User only need to input the species name (e.g. "Homo sapiens", "Mus musculus").
It will call \code{msigdbr} to download data from MSigDB and save it as RData under the \code{db/} directory with species name.
}
\examples{
gs.preload(use_spe='Homo sapiens',update=FALSE)
gs.preload(use_spe='Mus musculus',update=FALSE)
print(all_gs2gene_info)
# contain the information for all gene set category, category info, category size,
## sub category,sub category info, sub category size
print(names(all_gs2gene)) # the first level of the list is the category and sub-category IDs
print(str(all_gs2gene$`CP:KEGG`))

\dontrun{
gs.preload(use_spe='Homo sapiens',update=TRUE)
}
}
