% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{load.exp.GEO}
\alias{load.exp.GEO}
\title{Load gene expression set from GEO}
\usage{
load.exp.GEO(out.dir = network.par$out.dir.DATA, GSE = NULL,
  GPL = NULL, getGPL = TRUE, update = TRUE)
}
\description{
Load gene expression set from GEO
}
