% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{load.exp.RNASeq.demo}
\alias{load.exp.RNASeq.demo}
\title{Load Gene Expression Set from RNA-Seq Results (demo version)}
\usage{
load.exp.RNASeq.demo(files, type = "salmon", tx2gene = NULL,
  use_phenotype_info = NULL, use_sample_col = NULL,
  use_design_col = NULL, return_type = "tpm", merge_level = "gene")
}
\arguments{
\item{files}{a vector of characters, the filenames for the transcript-level abundances. It will be passed to \code{tximport}.
For details, please check \code{tximport}.}

\item{type}{character, the type of software used to generate the abundances. It will be passed to \code{tximport}.
For details, please check \code{tximport}.}

\item{tx2gene}{data.frame or NULL, this parameter will be passed to \code{tximport}. For details, please check \code{tximport}.}

\item{use_phenotype_info}{data.frame, the data frame contains phenotype information. It must have the columns \code{use_sample_col} and \code{use_design_col}.}

\item{use_sample_col}{character, the column name, indicating which column in \code{use_phenotype_info} should be used as the sample name.}

\item{use_design_col}{character, the column name, indicating which column in \code{use_phenotype_info} should be used as the design feature for samples.}

\item{return_type}{character, the class of the return object.
"txi" is the output of \code{tximport}. It is a list containing three matrices, abundance, counts and length.
"counts" is the matrix of raw count.
"tpm" is the raw tpm.
"fpm", "cpm" is the fragments/counts per million mapped fragments.
"raw-dds" is the DESeqDataSet class object, which is the original one without processing.
"dds" is the DESeqDataSet class object, which is processed by \code{DESeq}.
"eset" is the ExpressionSet class object, which is processed by \code{DESeq} and \code{vst}.
Default is "tpm".}

\item{merge_level}{character, users can choose between "gene" and "transcript".
"gene", the original salmon results will be mapped to the transcriptome and the expression matrix will be merged to the gene level.
This only works when using e.g. "gencode.vXX.transcripts.fa" from GENCODE as the reference.}
}
\description{
\code{load.exp.RNASeq.demo} is a function to read in RNA-Seq results and convert it to \code{eSet/DESeqDataSet} class object.
}
\details{
This function helps users to read in RNA-Seq results from various sources.
Due to the complicated manipulations (e.g. reference sequence) in processing RNA-Seq, this demo function may not be suitable for all scenarios.
}
