% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{merge.eset}
\alias{merge.eset}
\title{Merge two expression set.}
\usage{
\method{merge}{eset}(eset1, eset2, group1, group2, phe1, phe2,
  use_col = NULL, remove_batch = TRUE)
}
\description{
Merge two expression set.
}
