% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{merge.TF_SIG.AC}
\alias{merge.TF_SIG.AC}
\title{merge activity matrix for TF and SIG}
\usage{
\method{merge}{TF_SIG.AC}(TF_AC = NULL, SIG_AC = NULL)
}
\description{
merge activity matrix for TF and SIG
}
