% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{merge.TF_SIG.network}
\alias{merge.TF_SIG.network}
\title{merge target network for TF and SIG}
\usage{
\method{merge}{TF_SIG.network}(TF_network = NULL, SIG_network = NULL)
}
\description{
merge target network for TF and SIG
}
