% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{merge_eset}
\alias{merge_eset}
\title{Merge two expression set.}
\usage{
merge_eset(eset1,eset2,group1,group2,phe1,phe2,use_col = NULL,remove_batch = TRUE)
}
\description{
Merge two expression set.
}
