% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{merge_target_list}
\alias{merge_target_list}
\title{Merge target list for two drivers.}
\usage{
merge_target_list(driver1 = NULL, driver2 = NULL, target_list = NULL)
}
\arguments{
\item{driver1}{character, the name for the first driver to merge.}

\item{driver2}{character, the name for the second driver to merge.}

\item{target_list}{a list for the target gene information for the drivers. The names for the list must contain the driver1 and driver2.
Each object in the list must be a data.frame and should contain one column ("target") to save the target genes.
Strongly suggest to follow the NetBID2 pipeline, and the \code{target_list} could be automatically generated by \code{get_net2target_list} by
running \code{get.SJAracne.network}.}
}
\value{
a list for the target gene information.
}
\description{
\code{merge_target_list} is a function to merge the target list for two drivers.
Higher MI statistics for the shared target genes by the two drivers will be kept in the final target list.
}
\examples{
analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo1','driver/DATA/',package = "NetBID2")
NetBID.loadRData(analysis.par=analysis.par,step='ms-tab')
ms_tab <- analysis.par$final_ms_tab
driver1 <- ms_tab[1,'originalID_label']
driver2 <- ms_tab[2,'originalID_label']
m1 <- merge_target_list(driver1=driver1,driver2=driver2,
                          target_list=analysis.par$merge.network$target_list)
}
