% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{merge_TF_SIG.network}
\alias{merge_TF_SIG.network}
\title{Merge TF (transcription factor) Network and Sig (signaling factor) Network}
\usage{
merge_TF_SIG.network(TF_network = NULL, SIG_network = NULL)
}
\arguments{
\item{TF_network}{list, the TF network created by \code{get.SJAracne.network} function.}

\item{SIG_network}{list, the SIG network created by \code{get.SJAracne.network} function.}
}
\value{
Return the a list containing three elements, \code{network_dat}, \code{target_list} and \code{igraph_obj}.
}
\description{
\code{merge_TF_SIG.network} takes TF network and Sig network and combine them together.
The merged list object contains three elements, a data.frame contains all the combined network information \code{network_dat},
a driver-to-target list object \code{target_list}, and an igraph object of the network \code{igraph_obj}.
}
\examples{
if(exists('analysis.par')==TRUE) rm(analysis.par)
network.dir <- sprintf('\%s/demo1/network/',system.file(package = "NetBID2")) # use demo
network.project.name <- 'project_2019-02-14' # demo project name
project_main_dir <- 'test/'
project_name <- 'test_driver'
analysis.par  <- NetBID.analysis.dir.create(project_main_dir=project_main_dir,
                                            project_name=project_name,
                                            network_dir=network.dir,
                                            network_project_name=network.project.name)
analysis.par$tf.network  <- get.SJAracne.network(network_file=analysis.par$tf.network.file)
analysis.par$sig.network <- get.SJAracne.network(network_file=analysis.par$sig.network.file)
analysis.par$merge.network <- merge_TF_SIG.network(TF_network=analysis.par$tf.network,
                                                   SIG_network=analysis.par$sig.network)
}
