% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{NetBID.lazyMode.DriverVisualization}
\alias{NetBID.lazyMode.DriverVisualization}
\title{Lazy mode for NetBID2 result visualization}
\usage{
NetBID.lazyMode.DriverVisualization(analysis.par = NULL,
  intgroup = NULL, use_comp = NULL,
  main_id_type = "external_gene_name", transfer_tab = NULL,
  use_gs = c("H", "CP:BIOCARTA", "CP:REACTOME", "CP:KEGG"),
  min_Size = 30, max_Size = 1000, top_number = 30,
  top_strategy = "Both", logFC_thre = 0.05, Pv_thre = 0.05)
}
\arguments{
\item{analysis.par}{list, stores all related datasets from driver analysis step.}

\item{intgroup}{character, one interested phenotype group from the \code{analysis.par$cal.eset}.}

\item{use_comp}{character, the name of the comparison of interest, should be included in the colnames of \code{analysis.par$DA} and \code{analysis.par$DE}.}

\item{main_id_type}{character, the type of driver's ID. It comes from the attribute name in biomaRt package.
Such as "ensembl_gene_id", "ensembl_gene_id_version", "ensembl_transcript_id", "ensembl_transcript_id_version" or "refseq_mrna".
For details, user can call \code{biomaRt::listAttributes()} to display all available attributes in the selected dataset.}

\item{transfer_tab}{data.frame, the ID conversion table. Users can call \code{get_IDtransfer} to get this table.}

\item{use_gs}{a vector of characters, names of major gene set collections. Users can call \code{all_gs2gene_info} to see all the available collections.
Default is c("H", "CP:BIOCARTA", "CP:REACTOME", "CP:KEGG").}

\item{min_Size}{numeric, minimum size for the target genes. Default is 30.}

\item{max_Size}{numeric, maximum size for the target genes. Default is 1000.}

\item{top_number}{number for the top significant genes/drivers in the combine results to be displayed on the plot.
Default is 30.}

\item{top_strategy}{character, choose from "Both", "Up", "Down".
If set to "Both", top drivers with highest absolute Z statistics will be displayed.
If set to "Up", only top up-regulated drivers will be displayed.
If set to "Down", only top down-regulated drivers will be displayed.
Default is "Both".}

\item{logFC_thre}{numeric, the threshold of logFC. Genes or drivers with absolute logFC value higher than the threshold will be kept.
Default is 0.05.}

\item{Pv_thre}{numeric, the threshold of P-values. Genes or drivers with P-values lower than the threshold will be kept.
Default is 0.05.}
}
\description{
\code{NetBID.lazyMode.DriverVisualization} is an integrated function to draw visualization plots for top drivers.
}
\details{
User need to strictly follow the NetBID2 pipeline to get the complicated list object analysis.par.
"intgroup", "use_comp" should be specified;
"transfer_tab" could be set to NULL with "main_id_type" specified, but it is suggested to input by hand if available.
}
\examples{
\dontrun{
analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo1','driver/DATA/',package = "NetBID2")
NetBID.loadRData(analysis.par=analysis.par,step='ms-tab')
analysis.par$out.dir.PLOT <- 'test/'
NetBID.lazyMode.DriverVisualization(analysis.par=analysis.par,
                                    intgroup='subgroup',use_comp='G4.Vs.others',
                                    transfer_tab=analysis.par$transfer_tab,
                                    logFC_thre=0.2,Pv_thre=1e-4)
}
}
