% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{NetBID.loadRData}
\alias{NetBID.loadRData}
\title{Reload Saved RData Created by \code{NetBID.saveRData}.}
\usage{
NetBID.loadRData(network.par = NULL, analysis.par = NULL, step = "exp-load")
}
\arguments{
\item{network.par}{list, stores all related datasets from network construction step.}

\item{analysis.par}{list, stores all related datasets from driver analysis step.}

\item{step}{character, name of the pipeline step. It should be previously assigned by user when calling \code{NetBID.saveRData} function.}
}
\description{
\code{NetBID.loadRData} is a function loads RData saved by \code{NetBID.saveRData} function.
It prevents user from repeating former pipeline steps.
}
\examples{
analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo1','driver/DATA/',package = "NetBID2")
NetBID.loadRData(analysis.par=analysis.par,step='ms-tab')

}
