% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{NetBID.network.dir.create}
\alias{NetBID.network.dir.create}
\title{Manipulation of Working Directories for NetBID2 Network Construction Step}
\usage{
NetBID.network.dir.create(project_main_dir = NULL, project_name = NULL)
}
\arguments{
\item{project_main_dir}{character, name or absolute path of the main working directory.}

\item{project_name}{character, name of the project folder.}
}
\value{
\code{NetBID.network.dir.create} returns a list object, containing main.dir (path of the main working directory),
project.name (project name), out.dir (path of the project folder, which contains three subfolders), out.dir.QC,
out.dir.DATA and out.dir.SJAR.
}
\description{
\code{NetBID.network.dir.create} is used to help users create an organized working directory for the network construction step in NetBID2 analysis.
However, it is not essential for the analysis.
It creates a hierarchcial working directory and returns a list contains this directory information.
}
\details{
This function needs users to define the main working directory and the project's name.
It creates a main working directory with a subdirectory of the project.
It also automatically creates three subfolders (QC, DATA and SJAR) within the project folder. QC/,
storing Quality Control related plots; DATA/, saving data in RData format;
SJAR/, storing files needed for running SJAracne command.
This function also returns a list object (example, \code{network.par} in the demo) with directory information wrapped inside.
This list is an essential for
network construction step, all the important intermediate data generated later will be wrapped inside.
}
\examples{

\dontrun{
# Creating a main working directory under the current working directory by folder name
network.par <- NetBID.network.dir.create("MyMainDir","MyProject")
# Or creating a main working directory under the current working directory by relative path
network.par <- NetBID.network.dir.create("./MyMainDir","MyProject")
# Or creating a main working directory to a specific path by absolute path
network.par <- NetBID.network.dir.create("~/Desktop/MyMainDir","MyProject")
}
}
