% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetBID.R
\name{netbid}
\alias{netbid}
\title{Network-based Bayesian Inference of Drivers}
\usage{
netbid(gsc, eset, group.case, group.ctrl, signed = FALSE)
}
\arguments{
\item{gsc}{gsc (gene set collection) is a input network in list format with drivers as names and network targets as values that can be mapped to featureNames(eset)}

\item{eset}{ExpressionSet object of the data used to make inference: pData(eset)$group must exist to indicate the groups which you will compare to make the inference; featureNames(eset) are ids that can be mapped by gsc values.}

\item{group.case}{case name in pData(eset)$group}

\item{group.ctrl}{control name in pData(eset)$group}

\item{signed}{If TURE, consider directions of case vs. ctrl, If FALSE, ignore the directions}
}
\value{
data.frame of case vs. ctrl inference with statistics including netSize (network size), z, pval, FDR.BH (FDR by BH method), log2FC (if signed)
}
\description{
Driver inference of of case phenotype (vs. ctrl) from one source data and a pre-defined network
}
\examples{

###transform network into list of probes
dci.kinase.probes<-network.tbl2list(dci.kinase.tbl[,c('source.symbol','target')])
dmrna<-netbid(dci.kinase.probes,mrna,group.case='CD8pos.DC',group.ctrl='CD8neg.DC')



}
