% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{NetBID2.loadRData}
\alias{NetBID2.loadRData}
\title{Automatically load RData for NetBID2.}
\usage{
NetBID2.loadRData(network.par = NULL, analysis.par = NULL,
  step = "exp-load")
}
\description{
Automatically load RData for NetBID2.
}
