% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{NetBID2.saveRData}
\alias{NetBID2.saveRData}
\title{Automatically save RData for NetBID2.}
\usage{
NetBID2.saveRData(network.par = NULL, analysis.par = NULL,
  step = "exp-load")
}
\description{
Automatically save RData for NetBID2.
}
