% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetBID.R
\name{netbidi}
\alias{netbidi}
\title{NetBID Integration}
\usage{
netbidi(netbid.list, signed = FALSE)
}
\arguments{
\item{netbid.list}{list of netbid outputs, e.g. list(mRNA=dmrna,wProtein=dwprotein,pProtein=dpprotein)}

\item{signed}{If TURE, consider directions when combining NetBID results, If FALSE, ignore the directions}
}
\value{
data.frame of case vs. ctrl inference with statistics including cobmbined and individual netSize (network size), z, pval, FDR.BH (FDR by BH method), log2FC (if signed)
}
\description{
Driver inference by intergrating multiple NetBID results
}
\examples{

###transform network into list of probes
dci.kinase.probes<-network.tbl2list(dci.kinase.tbl[,c('source.symbol','target')])
dmrna<-netbid(dci.kinase.probes,mrna,group.case='CD8pos.DC',group.ctrl='CD8neg.DC')
##whole proteomcis
dci.kinase.wprotein<-transformNetworkListIds(dci.kinase.symbols,ref=fData(wprotein)[,c('geneSymbol','proteinId')])
dwprotein<-netbid(dci.kinase.wprotein,wprotein,group.case='CD8pos.DC',group.ctrl='CD8neg.DC')
dcomb<-netbidi(list(mRNA=dmrna,wProtein=dwprotein))



}
