% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network.tbl2list}
\alias{network.tbl2list}
\title{Create network list format from table}
\usage{
network.tbl2list(net, annotation = NULL, annotateSize = FALSE)
}
\arguments{
\item{net}{network in tbl format, the first two columns: source, target}

\item{annotation}{if not NULL, it has two columns: the first column is ids that are mapped to source and target columns; the second column is the annotation such as geneSymbol, etc}

\item{annotateSize}{if TRUE, add size of the network}
}
\value{
network in list format
}
\description{
Transform network from table format to list format
}
\examples{

dci.kinase.probes<-network.tbl2list(dci.kinase.tbl[,c('source.symbol','target')])

}
