% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{out2excel}
\alias{out2excel}
\title{Output master table to excel files.}
\usage{
out2excel(all_ms_tab, out.xlsx, mark_gene = NULL, mark_col = NULL,
  mark_strategy = "color", workbook_name = "ms_tab",
  only_z_sheet = FALSE, z_column = NULL, sig_thre = 1.64)
}
\arguments{
\item{all_ms_tab}{list or dataframe, if dataframe, it is the master table generated by \code{generate.masterTable}.
If it is a list, each item in list could be a dataframe containing the master table.
The name of the list will be the sheet name in the excel file.}

\item{out.xlsx}{character, file name for the output excel.}

\item{mark_gene}{list, list of marker genes, additional info to add in the master table.The name of the list is the name for the mark group.}

\item{mark_col}{character, the color used to label the marker genes. If NULL, will use \code{get.class.color} to get the colors.}

\item{mark_strategy}{character, choose from 'color','add_column'. 'Color' means the mark_gene will be displayed by its background color;
'add_column' means the mark_gene will be displayed in separate columns with content TRUE/FALSE indicating whether the genes belong to each mark group.}

\item{workbook_name}{character, workbook name for the output excel. Default is 'ms_tab'.}

\item{only_z_sheet}{logical, if TRUE will generate a separate sheet only contain Z related columns in DA/DE. Default is FALSE.}

\item{z_column}{character, the column name that contain the Z-statistics. If NULL, will find columns start with "Z.".
Default is NULL.}

\item{sig_thre}{numeric, threshold for the Z-statistics, values passed the threshold will be marked by the color automatically generated by \code{z2col}.
Default is 1.64.}
}
\value{
logical value indicating whether the file has been sucessfully generated or not.
}
\description{
\code{out2excel} is a function to output dataframe into an excel file, mainly for output the master table generated by \code{generate.masterTable}.
}
\examples{
\dontrun{
analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo1','driver/DATA/',package = "NetBID2")
NetBID.loadRData(analysis.par=analysis.par,step='ms-tab')
ms_tab <- analysis.par$final_ms_tab ## this is master table generated before
mark_gene <- list(WNT=c('WIF1','TNC','GAD1','DKK2','EMX2'),
                 SHH=c('PDLIM3','EYA1','HHIP','ATOH1','SFRP1'),
                 Group3=c('IMPG2','GABRA5','EGFL11','NRL','MAB21L2','NPR3','MYC'),
                 Group4=c('KCNA1','EOMES','KHDRBS2','RBM24','UNC5D'))
mark_col <- get.class.color(names(mark_gene))
outfile <- 'test_out.xlsx'
out2excel(ms_tab,out.xlsx = out_file,mark_gene,mark_col)
}
}
