% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{plot.2D.MICA}
\alias{plot.2D.MICA}
\title{MICA in 2D}
\usage{
plot.2D.MICA(outdir = NULL, prjname = NULL, all_k = NULL,
  obs_label = NULL, legend_pos = "topleft", legend_cex = 0.8,
  point_cex = 1, plot_type = "2D.ellipse",
  visualization_type = "tsne")
}
\arguments{
\item{obs_label}{the value should be the class label with names equal to sample names}

\item{plot_type}{2D or 2D.ellipse}

\item{visualization_type}{tsne or umap}
}
\description{
MICA in 2D
}
