% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{plot.2D.pca.kmeans}
\alias{plot.2D.pca.kmeans}
\title{pca+kmeans in 2D}
\usage{
plot.2D.pca.kmeans(mat = NULL, all_k = NULL, obs_label = NULL,
  legend_pos = "topleft", legend_cex = 0.8, plot_type = "2D.ellipse",
  point_cex = 1)
}
\arguments{
\item{obs_label}{the value should be the class label with names equal to sample names}

\item{plot_type}{2D or 2D.ellipse}
}
\description{
pca+kmeans in 2D
}
