% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{plot.3D.MICA}
\alias{plot.3D.MICA}
\title{MICA in 3D}
\usage{
plot.3D.MICA(outdir = NULL, prjname = NULL, all_k = NULL,
  obs_label = NULL, legend_pos = "topleft", legend_cex = 0.8,
  legend_ncol = 1, point_cex = 1, visualization_type = "umap")
}
\arguments{
\item{obs_label}{the value should be the class label with names equal to sample names}

\item{visualization_type}{tsne or umap}

\item{visualization_type}{mds,umap}
}
\description{
MICA in 3D
}
