% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{plot.3D.pca.kmeans}
\alias{plot.3D.pca.kmeans}
\title{pca+kmeans in 3D}
\usage{
plot.3D.pca.kmeans(mat = NULL, all_k = NULL, obs_label = NULL,
  legend_pos = "topleft", legend_ncol = 1, legend_cex = 0.8)
}
\arguments{
\item{obs_label}{the value should be the class label with names equal to sample names}
}
\description{
pca+kmeans in 3D
}
