% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{plot.3D.umap.kmeans}
\alias{plot.3D.umap.kmeans}
\title{umap+kmeans in 3D}
\usage{
plot.3D.umap.kmeans(mat = NULL, all_k = NULL, obs_label = NULL,
  legend_pos = "topleft", legend_cex = 0.8, point_cex = 1)
}
\arguments{
\item{obs_label}{the value should be the class label with names equal to sample names}
}
\description{
umap+kmeans in 3D
}
