% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{plot.eset.QC}
\alias{plot.eset.QC}
\title{QC plot for eset}
\usage{
plot.eset.QC(eset, outdir = ".", do.logtransform = FALSE,
  intgroup = NULL, prefix = "afterQC_", choose_plot = c("heatmap",
  "pca", "density", "meansd"))
}
\arguments{
\item{eset}{input}

\item{outdir}{output directory}

\item{intgroup}{interested group}
}
\description{
QC plot for eset
}
