% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{RNASeqCount.normalize.scale}
\alias{RNASeqCount.normalize.scale}
\title{Normalization of RNA-Seq Reads Count}
\usage{
RNASeqCount.normalize.scale(mat, total = NULL, pseudoCount = 1)
}
\arguments{
\item{mat}{matrix, matrix of RNA-Seq reads data. Each row is a gene/transcript, each column is a sample.}

\item{total}{integer, total RNA-Seq reads count. If NULL, will use the mean of each column's summation. Default is NULL.}

\item{pseudoCount}{integer, the integer added to avoid "-Inf" showing up during log transformation. Default is 1.}
}
\value{
Return a numeric matrix, containing the normalized RNA-Seq reads count.
}
\description{
\code{RNASeqCount.normalize.scale} is a simple version to normalize the RNASeq reads count data.
}
\details{
Users can also load \code{load.exp.RNASeq.demo}, and follow the \code{DESeq2} pipeline for RNASeq data processing.
Warning, \code{load.exp.RNASeq.demo} and \code{load.exp.RNASeq.demoSalmon} in NetBID2 may not cover all the possible scenarios.
}
\examples{
mat1 <- matrix(rnbinom(10000, mu = 10, size = 1),nrow=1000,ncol=10)
colnames(mat1) <- paste0('Sample1',1:ncol(mat1))
rownames(mat1) <- paste0('Gene',1:nrow(mat1))
norm_mat1 <- RNASeqCount.normalize.scale(mat1)
}
