% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{SJ.MICA.prepare}
\alias{SJ.MICA.prepare}
\title{Inner use: prepare for MICA}
\usage{
SJ.MICA.prepare(mat, outdir = ".", prjname = NULL, all_k = NULL,
  retransformation = "False", perplexity = 30)
}
\description{
Inner use: prepare for MICA
}
