% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{SJ.MICA.prepare.withfile}
\alias{SJ.MICA.prepare.withfile}
\title{Inner use: prepare for MICA with file}
\usage{
SJ.MICA.prepare.withfile(input_exp, outdir = ".", prjname = NULL,
  all_k = NULL, retransformation = "False", perplexity = 30)
}
\description{
Inner use: prepare for MICA with file
}
