% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{SJ.SJAracne.prepare}
\alias{SJ.SJAracne.prepare}
\title{Inner use: prepare for SJaracne run on SJ server
prepare SJAracne dataset for net-dataset
project_name expression_matrix hub_genes outdir}
\usage{
SJ.SJAracne.prepare(eset, use.samples = rownames(pData(eset)),
  TF_list = NULL, SIG_list = NULL, SJAR.project_name = "",
  SJAR.main_dir = NULL, SJAR.bash.main_dir = NULL, mem = 40960,
  IQR.thre = 0.5, IQR.loose_thre = 0.1)
}
\description{
Inner use: prepare for SJaracne run on SJ server
prepare SJAracne dataset for net-dataset
project_name expression_matrix hub_genes outdir
}
