% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{SJ.SJAracne.step}
\alias{SJ.SJAracne.step}
\title{Inner use: auto generate bash files for Step1, Step2, Step3, Step4 for all bash files under one directory}
\usage{
SJ.SJAracne.step(out.dir.SJAR)
}
\description{
Inner use: auto generate bash files for Step1, Step2, Step3, Step4 for all bash files under one directory
}
