% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{SJAracne.prepare}
\alias{SJAracne.prepare}
\title{Prepare for running SJaracne}
\usage{
SJAracne.prepare(eset, use.samples = rownames(pData(eset)),
  TF_list = NULL, SIG_list = NULL, SJAR.main_dir = "",
  SJAR.project_name = "", IQR.thre = 0.5, IQR.loose_thre = 0.1,
  add_options = "")
}
\arguments{
\item{eset}{an ExpressionSet class object, which contains the expression matrix.}

\item{use.samples}{a vector of characters, the sample list used for running SJAracne.}

\item{TF_list}{a vector of characters, the TF list used for analysis.}

\item{SIG_list}{a vector of characters, the SIG list used for analysis.}

\item{SJAR.main_dir}{character, the file path to save the results for SJAracne}

\item{SJAR.project_name}{character, the name of the project used to label the output directory.}

\item{IQR.thre}{numeric, threshold for IQR filter for all non-driver genes.}

\item{IQR.loose_thre}{numeric, threshold for IQR filter for all driver(TF/SIG) genes.}

\item{add_options}{additional options used to run sjaracne.}
}
\description{
\code{SJAracne.prepare} is a function to prepare files for running SJAracne.
}
\details{
Detailed description to run SJAracne could be found in Github.
Check \url{https://github.com/jyyulab/SJARACNe/} for detail.
}
\examples{
\dontrun{
network.par <- list()
network.par$out.dir.DATA <- system.file('demo1','network/DATA/',package = "NetBID2")
NetBID.loadRData(network.par=network.par,step='exp-QC')
db.preload(use_level='gene',use_spe='human',update=FALSE)
use_gene_type <- 'external_gene_name' ## this should user-defined !!!
use_genes <- rownames(fData(network.par$net.eset))
use_list  <- get.TF_SIG.list(use_genes,use_gene_type=use_gene_type)
#select sample for analysis
phe <- pData(network.par$net.eset)
use.samples <- rownames(phe) ## use all samples, or choose to use some samples
prj.name <- network.par$project.name # can use other names, if need to run different use samples
network.par$out.dir.SJAR <- 'test' ## set the directory
SJAracne.prepare(eset=network.par$net.eset,
                 use.samples=use.samples,
                 TF_list=use_list$tf,
                 SIG_list=use_list$sig,
                 IQR.thre = 0.5,IQR.loose_thre = 0.1,
                 SJAR.project_name=prj.name,
                 SJAR.main_dir=network.par$out.dir.SJAR)
}
}
