% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{test.targetNet.overlap}
\alias{test.targetNet.overlap}
\title{Test for the target genes' intersection between two drivers.}
\usage{
test.targetNet.overlap(source1_label = NULL, source2_label = NULL,
  target1 = NULL, target2 = NULL, total_possible_target = NULL)
}
\arguments{
\item{source1_label}{character, the label for the first driver.}

\item{source2_label}{character, the label for the second driver.}

\item{target1}{a vector of characters, the list of target genes for the first driver.}

\item{target2}{a vector of characters, the list of target genes for the second driver.}

\item{total_possible_target}{numeric or a vector of characters. If input numeric, will be the total number of possible targets.
If input a vector of characters, will be the background list of all possible target genes.}
}
\value{
Return the statistics of testing, including the \code{P.Value}, \code{Odds_Ratio} and \code{Intersected_Nuumber}.
}
\description{
\code{test.targetNet.overlap} will test whether the target genes of two drivers are significantly intersected.
}
\details{
This is a function to perform Fisher's Exact Test for the intersection of the target genes from two drivers.
}
\examples{
source1_label <- 'test1'
target1 <- sample(paste0('G',1:1000),size=80)
source2_label <- 'test2'
target2 <- sample(paste0('G',1:1000),size=120)
test.targetNet.overlap(source1_label=source1_label,source2_label=source2_label,
               target1=target1,target2=target2,
               total_possible_target=paste0('G',1:1000))
\dontrun{
}
}
