% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{transformNetworkListIds}
\alias{transformNetworkListIds}
\title{Transform network ids}
\usage{
transformNetworkListIds(net.list, ref)
}
\arguments{
\item{net.list}{network in list format}

\item{ref}{data.frame with two columns: id in net.list, new id to tranform}
}
\value{
network in list format
}
\description{
Transform network ids given a reference, e.g. from gene symbols to probe ids
}
\examples{

dci.kinase.wprotein<-transformNetworkListIds(dci.kinase.symbols,ref=fData(wprotein)[,c('geneSymbol','proteinId')])

}
