% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{update.eset.feature}
\alias{update.eset.feature}
\title{Update eset feature, mainly for gene ID conversion}
\usage{
update.eset.feature(use_eset = NULL, use_feature_info = NULL,
  from_feature = NULL, to_feature = NULL, merge_method = "median")
}
\description{
ID conversion, or merge transcript level to expression level, use_feature_info can be other dataframe info
merge_method: mean median max min
if multi-from --> one to: merge
if one-from --> multi to: equal distribute.
}
