% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{update.eset.phenotype}
\alias{update.eset.phenotype}
\title{Update eset phenotype}
\usage{
update.eset.phenotype(use_eset = NULL, use_phenotype_info = NULL,
  use_sample_col = NULL, use_col = "GEO-auto")
}
\description{
Update eset phenotype
}
