% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{update_eset.phenotype}
\alias{update_eset.phenotype}
\title{Update ExpressionSet phenotype information, mainly for changing sample name and extract useful phenotype information for sample clustering.}
\usage{
update_eset.phenotype(use_eset = NULL, use_phenotype_info = NULL,
  use_sample_col = NULL, use_col = "auto")
}
\arguments{
\item{use_eset}{ExpressionSet class object, the original ExpressionSet to update.}

\item{use_phenotype_info}{data.frame, the phenotype information dataframe.}

\item{use_sample_col}{character,the column name in \code{use_phenotype_info}, which contains the sample name.}

\item{use_col}{character, the columns in \code{use_phenotype_info} to be kept.
If set to 'auto', wil extract columns with unique sample feature ranges from 2 to sample size-1.
If set to 'GEO-auto', will extract columns: 'geo_accession','title','source_name_ch1',and columns end with ':ch1'.
Default is "auto".}
}
\value{
an ExressionSet object.
}
\description{
\code{update_eset.phenotype} is a function to update the phenotype information in the ExpressionSet object.
}
\details{
This function is designed mainly for extracting useful phenotype information for samples.
Especially for phenotype information directly get from GEO database.
}
\examples{
\dontrun{
net_eset <- load.exp.GEO(out.dir='test/',
                         GSE='GSE116028',
                         GPL='GPL6480',
                         getGPL=TRUE,
                         update=FALSE)
net_eset <- update_eset.phenotype(use_eset=net_eset,
                                  use_phenotype_info=pData(net_eset),
                                  use_sample_col='geo_accession',
                                  use_col='GEO-auto')
}
}
