% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{cal.Activity.GS}
\alias{cal.Activity.GS}
\title{Calculate Activity Value for Gene Sets}
\usage{
cal.Activity.GS(use_gs2gene = all_gs2gene[c("H", "CP:BIOCARTA",
  "CP:REACTOME", "CP:KEGG")], cal_mat = NULL, es.method = "mean",
  std = TRUE)
}
\arguments{
\item{use_gs2gene}{list, contains elements of gene sets. Element name is gene set name, each element contains a vector of genes belong to that gene set.
Default is using \code{all_gs2gene[c('H','CP:BIOCARTA','CP:REACTOME','CP:KEGG')]}, which is loaded from \code{gs.preload}.}

\item{cal_mat}{numeric matrix, gene/transcript expression matrix.}

\item{es.method}{character, method to calculate the activity value. Users can choose from "mean", "absmean", "maxmean", "gsva", "ssgsea", "zscore" and "plage".
The details for using the last four options, users can check \code{gsva}. Default is "mean".}

\item{std}{logical, if TRUE, the expression matrix will be normalized by column. Default is TRUE.}
}
\value{
Return an activity matrix with rows of gene sets and columns of samples.
}
\description{
\code{cal.Activity.GS} calculates activity value for each gene set, and return a numeric matrix with rows of gene sets and columns of samples.
}
\examples{
analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo1','driver/DATA/',package = "NetBID2")
NetBID.loadRData(analysis.par=analysis.par,step='ms-tab')
gs.preload(use_spe='Homo sapiens',update=FALSE)
use_gs2gene <- merge_gs(all_gs2gene=all_gs2gene,
                        use_gs=c('H','CP:BIOCARTA','CP:REACTOME','CP:KEGG','C5'))
exp_mat_gene <- Biobase::exprs(analysis.par$cal.eset)
## each row is a gene symbol, if not, must convert ID first
ac_gs <- cal.Activity.GS(use_gs2gene = use_gs2gene,
                        cal_mat = exp_mat_gene)
}
