% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{cal.Activity.GS}
\alias{cal.Activity.GS}
\title{calculate activity for gene sets}
\usage{
cal.Activity.GS(use_gs2gene = all_gs2gene[c("H", "CP:BIOCARTA",
  "CP:REACTOME", "CP:KEGG")], cal_mat = NULL, es.method = "mean")
}
\arguments{
\item{es.method:}{mean, absmean, maxmean}
}
\description{
calculate activity for gene sets
}
