% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{cal.Activity}
\alias{cal.Activity}
\title{Calculate Activity Value for Each Driver}
\usage{
cal.Activity(target_list = NULL, cal_mat = NULL,
  es.method = "weightedmean", std = TRUE)
}
\arguments{
\item{target_list}{list, the driver-to-target list object. The names of the list elements are drivers.
Each element is a data frame, usually contains at least three columns. "target", target gene names;
"MI", mutual information; "spearman", spearman correlation coefficient. Users can call \code{get_net2target_list} to create this list from the network file generated by SJAracne.}

\item{cal_mat}{numeric matrix, the expression matrix of genes/transcripts.}

\item{es.method}{character, method applied to calculate the activity value. User can choose from "mean", "weightedmean", "maxmean" and "absmean".
The "weightedmean" requires "MI" and "spearman" columns included in the \code{target_list} data.
Default is 'weightedmean'.}

\item{std}{logical, if TRUE, the expression matrix will be normalized by column. Default is TRUE.}
}
\value{
Return a matrix of activity values. Rows are drivers, columns are samples.
}
\description{
\code{cal.Activity} calculates the activity value for each driver.
This function requires two inputs, the driver-to-target list object \code{target_list} and the expression matrix.
}
\examples{
analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo1','driver/DATA/',package = "NetBID2")
NetBID.loadRData(analysis.par=analysis.par,step='ms-tab')
ac_mat <- cal.Activity(target_list=analysis.par$tf.network$target_list,
                       cal_mat=exprs(analysis.par$cal.eset),
                       es.method='weightedmean')
}
