% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{combinePvalVector}
\alias{combinePvalVector}
\title{Combine Pvalues by Fisher or Stouffer's method}
\usage{
combinePvalVector(pvals, method = c("Stouffer", "Fisher"),
  signed = TRUE, twosided = TRUE)
}
\arguments{
\item{pvals, }{a vector of numeric values, the P-value values need to combine.
The sign of the P-value will be added to indicate the direction of testing if signed is set to TRUE.}

\item{method}{character, choose from Stouffer or Fisher, default is "Stouffer".}

\item{signed}{logical, whether the sign of the pvals will be considered in calculation.
Default is TRUE.}

\item{twosided}{logical, whether the pvalues are from two-sided test or not.
If not, pvalues must between 0 and 0.5.
Default is TRUE.}
}
\value{
a vector contains the 'Z-statistics' and 'P.Value'
}
\description{
\code{combinePvalVector} is a function to combine P-values by Stouffer of Fisher method.
}
\examples{
combinePvalVector(c(0.1,1e-3,1e-5))
combinePvalVector(c(0.1,1e-3,-1e-5))
}
