% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{draw.2D.ellipse}
\alias{draw.2D.ellipse}
\title{Visualize Sample Clustering Result in 2D Plot with Ellipse}
\usage{
draw.2D.ellipse(X, Y, class_label, xlab = "PC1", ylab = "PC2",
  legend_cex = 0.8, main = "", point_cex = 1)
}
\arguments{
\item{X}{a vector of numerics, the x coordinates of points in the plot. If user would like to creat a PCA biplot, this parameter should be the first component.}

\item{Y}{a vector of numerics, the y coordinates of points in the plot. If user would like to creat a PCA biplot, this parameter should be the second component.}

\item{class_label}{a vector of characters, labels or categories of samples. The vector name should be sample names.}

\item{xlab}{character, the label for x-axis. Default is "PC1".}

\item{ylab}{character, the label for y-axis. Default is "PC2".}

\item{legend_cex}{numeric, giving the amount by which the text of legend should be magnified relative to the default. Default is 0.8.}

\item{main}{character, an overall title for the plot. Default is "".}

\item{point_cex}{numeric, giving the amount by which the size of the data points should be magnified relative to the default. Default is 1.}
}
\value{
Return a logical value. If TRUE, the plot has been created successfully.
}
\description{
\code{draw.2D.ellipse} creates a 2D plot with an ellipse drawn around each cluster to visualize the sample clustering result.
}
\examples{
mat1 <- matrix(rnorm(2000,mean=0,sd=1),nrow=100,ncol=20)
rownames(mat1) <- paste0('Gene',1:nrow(mat1))
colnames(mat1) <- paste0('Sample',1:ncol(mat1))
pc <- prcomp(t(mat1))$x
pred_label <- kmeans(pc,centers=4)$cluster ## this can use other cluster results
draw.2D.ellipse(X=pc[,1],Y=pc[,2],class_label=pred_label)
}
