% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{draw.3D}
\alias{draw.3D}
\title{Draw 3D dimension plot for sample cluster visualization.}
\usage{
draw.3D(X, Y, Z, class_label, xlab = "PC1", ylab = "PC2",
  zlab = "PC3", legend_cex = 0.8, main = "", point_cex = 1,
  legend_pos = "topright", legend_ncol = 1, ...)
}
\arguments{
\item{X}{a vector of numeric values, the first dimension values.}

\item{Y}{a vector of numeric values, the second dimension values.}

\item{Z}{a vector of numeric values, the third dimension values.}

\item{class_label}{a vector of characters, with the names are the samples (optional) and the values are the sample cluster label.
The function will treat that the order are the same for X,Y and class_label.}

\item{xlab}{character, the label for x-axis.}

\item{ylab}{character, the label for y-axis.}

\item{zlab}{character, the label for z-axis.}

\item{legend_cex}{numeric, the cex for the legend.}

\item{main}{character, the title for the plot.}

\item{point_cex}{numeric, the cex for the points.}

\item{legend_pos}{character, the position to put the legend. Default is 'topright'.}

\item{legend_ncol}{integer, number of columns used to display the legend. Default is 1.}

\item{...}{other paramters used in \code{scatter3D}.}
}
\value{
logical value indicating whether the plot has been successfully generated
}
\description{
\code{draw.3D} is a function to draw 3D dimension plot for sample cluster visualization.
}
\examples{
mat1 <- matrix(rnorm(2000,mean=0,sd=1),nrow=100,ncol=20)
rownames(mat1) <- paste0('Gene',1:nrow(mat1))
colnames(mat1) <- paste0('Sample',1:ncol(mat1))
pc <- prcomp(t(mat1))$x
pred_label <- kmeans(pc,centers=4)$cluster ## this can use other cluster results
draw.3D(X=pc[,1],Y=pc[,2],Z=pc[,3],class_label=pred_label)
}
