% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{draw.eset.QC}
\alias{draw.eset.QC}
\title{QC plots for ExpressionSet class object.}
\usage{
draw.eset.QC(eset, outdir = ".", do.logtransform = FALSE,
  intgroup = NULL, prefix = "", choose_plot = c("heatmap", "pca",
  "density", "meansd"), generate_html = TRUE)
}
\arguments{
\item{eset}{ExpressionSet class, quality control analysis target.}

\item{outdir}{character, the directory path for saving output files.}

\item{do.logtransform}{logical, if TRUE, the log transformation will be performed on the gene expression value. Default is FALSE.}

\item{intgroup}{a vector of characters, the interested phenotype groups from the ExpressionSet.
If NULL, it will automatcially extract all possible groups by \code{get_int_group}.
Default is NULL.}

\item{prefix}{character, the prefix for the QC figures' name. Default is "".}

\item{choose_plot}{a vector of characters,
choose one or many from 'heatmap', 'pca', 'density' and 'meansd' plots.
Default is all.}

\item{generate_html}{logical, if TRUE, it will generate a html file by R Markdown. Otherwise, it will generate separate PDF files.
Default is TRUE.}
}
\description{
\code{draw.eset.QC} is a function to draw a set of plots for quality control analysis.
4 types of plots will be created, including heatmap, pca, density and meansd.
}
\examples{
\dontrun{
network.par <- list()
network.par$out.dir.DATA <- system.file('demo1','network/DATA/',package = "NetBID2")
NetBID.loadRData(network.par=network.par,step='exp-QC')
intgroups <- get_int_group(network.par$net.eset)
network.par$out.dir.QC <- getwd() ## set the output directory
draw.eset.QC(network.par$net.eset,outdir=network.par$out.dir.QC,intgroup=intgroups)
}
}
