% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{draw.funcEnrich.bar}
\alias{draw.funcEnrich.bar}
\title{Bar plot for the result of gene set enrichment analysis.}
\usage{
draw.funcEnrich.bar(funcEnrich_res = NULL, top_number = 30,
  Pv_col = "Ori_P", item_col = "Intersected_items", Pv_thre = 0.1,
  display_genes = FALSE, name_col = "#Name", gs_cex = 0.5,
  gene_cex = 0.5, main = "", bar_col = brewer.pal(8, "RdBu")[7],
  eg_num = 5, pdf_file = NULL)
}
\arguments{
\item{funcEnrich_res}{data.frame, containing the result for the function enrichment analysis. Prefer the format generated by using \code{funcEnrich.Fisher}.
If not, users could prepare the required columns and indicate the column names in the following parameters.}

\item{top_number}{numeric, number for the top enriched gene sets to be displayed on the plot. Default is 30.}

\item{Pv_col}{character, the name of the column in \code{funcEnrich_res}, which contains the P-value. Default is 'Ori_P'.}

\item{item_col}{character, the name of the column in \code{funcEnrich_res}, which contains the detailed intersected gene list, collapsed by ';'.
Default is 'Intersected_items'.}

\item{Pv_thre}{numeric, cutoff value for the p-values. Genes or drivers with lower p-values are remained. Default is 0.1.}

\item{display_genes}{logical, whether or not to display the intersected genes on the plot. Default is FALSE}

\item{name_col}{character, the name of the column in \code{funcEnrich_res}, which contains the gene set name. Default is '#Name'.}

\item{gs_cex}{numeric, \code{cex} for the gene sets displayed on the plot. Default is 0.5}

\item{gene_cex}{numeric, \code{cex} for the genes displayed on the plot. Default is 0.5}

\item{main}{character, \code{main} for the title on the plot.}

\item{bar_col}{character, color code for the bar on the plot. Default is brewer.pal(8,'RdBu')[7].}

\item{eg_num}{numeric, example number of intersected genes shown on the plot. Default is 5.}

\item{pdf_file}{character, file path for the pdf file to save the figure into pdf format.If NULL, will not generate pdf file. Default is NULL.}
}
\value{
logical value indicating whether the plot has been successfully generated
}
\description{
\code{draw.funcEnrich.bar} will draw the barplot for the result of gene set enrichment analysis.
}
\details{
This is a function to draw the barplot for the result of gene set enrichment analysis.
Two modes, one just display the P-value, and the other could show the top intersected genes for each gene set.
}
\examples{
analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo1','driver/DATA/',package = "NetBID2")
NetBID.loadRData(analysis.par=analysis.par,step='ms-tab')
ms_tab <- analysis.par$final_ms_tab
sig_driver <- draw.volcanoPlot(dat=ms_tab,label_col='gene_label',
                               logFC_col='logFC.G4.Vs.others_DA',
                               Pv_col='P.Value.G4.Vs.others_DA',
                               logFC_thre=0.4,
                               Pv_thre=1e-7,
                               main='Volcano Plot for G4.Vs.others_DA',
                               show_label=FALSE,
                               label_type = 'origin',
                               label_cex = 0.5)
gs.preload(use_spe='Homo sapiens',update=FALSE)
res1 <- funcEnrich.Fisher(input_list=ms_tab[rownames(sig_driver),'geneSymbol'],
                          bg_list=ms_tab[,'geneSymbol'],
                          use_gs=c('H','C5'),Pv_thre=0.1,
                          Pv_adj = 'none')
draw.funcEnrich.bar(funcEnrich_res=res1,top_number=5,
                   main='Function Enrichment for Top drivers',
                   gs_cex=0.4,gene_cex=0.5)
draw.funcEnrich.bar(funcEnrich_res=res1,top_number=3,
                   main='Function Enrichment for Top drivers',
                   display_genes = TRUE,eg_num=3,
                   gs_cex=0.3,gene_cex=0.3)
\dontrun{
analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo1','driver/DATA/',package = "NetBID2")
NetBID.loadRData(analysis.par=analysis.par,step='ms-tab')
ms_tab <- analysis.par$final_ms_tab
sig_driver <- draw.volcanoPlot(dat=ms_tab,label_col='gene_label',
                               logFC_col='logFC.G4.Vs.others_DA',
                               Pv_col='P.Value.G4.Vs.others_DA',
                               logFC_thre=0.4,
                               Pv_thre=1e-7,
                               main='Volcano Plot for G4.Vs.others_DA',
                               show_label=FALSE,
                               label_type = 'origin',
                               label_cex = 0.5)
gs.preload(use_spe='Homo sapiens',update=FALSE)
res1 <- funcEnrich.Fisher(input_list=ms_tab[rownames(sig_driver),'geneSymbol'],
                          bg_list=ms_tab[,'geneSymbol'],
                          use_gs=c('H','C5'),Pv_thre=0.1,Pv_adj = 'none')
analysis.par$out.dir.PLOT <- getwd() ## directory for saving the pdf files
draw.funcEnrich.bar(funcEnrich_res=res1,top_number=30,
                    main='Function Enrichment for Top drivers',
                    pdf_file=sprintf('\%s/funcEnrich_bar_nogene.pdf',
                    analysis.par$out.dir.PLOT))
draw.funcEnrich.bar(funcEnrich_res=res1,top_number=30,
                    main='Function Enrichment for Top drivers',
                    display_genes = TRUE,gs_cex=0.6,
                    pdf_file=sprintf('\%s/funcEnrich_bar_withgene.pdf',
                    analysis.par$out.dir.PLOT))
}
}
