% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{draw.GSEA.NetBID.GS}
\alias{draw.GSEA.NetBID.GS}
\title{Draw GSEA (gene set enrichment analysis) Plot with NetBID Analysis of Gene Sets}
\usage{
draw.GSEA.NetBID.GS(
  DE = NULL,
  name_col = NULL,
  profile_col = NULL,
  profile_trend = "pos2neg",
  sig_gs_list = NULL,
  gs_DA_Z = NULL,
  use_gs2gene = NULL,
  top_gs_number = 30,
  target_col = "RdBu",
  left_annotation = "",
  right_annotation = "",
  main = "",
  profile_sig_thre = 0,
  Z_sig_thre = 1.64,
  pdf_file = NULL
)
}
\arguments{
\item{DE}{data.frame, a data.frame created either by function \code{getDE.limma.2G} or \code{getDE.BID.2G}.
Row names are gene names, columns must include the calculated differencial values (e.g. "ID", "logFC", "AveExpr", "P.Value" etc.).}

\item{name_col}{character, the name of the column in \code{DE} contains gene names. If NULL, will use the row names of \code{DE}. Default is NULL.}

\item{profile_col}{character, the name of the column in \code{DE} contains calculated differencial value (e.g. "logFC" or "P.Value").
If \code{DE} is created by \code{getDE.limma.2G} or \code{getDE.BID.2G}, this parameter should be set to "logFC" or "t".}

\item{profile_trend}{character, users can choose between "pos2neg" and "neg2pos". "pos2neg" means high \code{profile_col} in target group will be shown on the left.
"neg2pos" means high \code{profile_col} in control group will be shown on the left. Default is "pos2neg".}

\item{sig_gs_list}{a vector of characters, the names of top gene sets.}

\item{gs_DA_Z}{a vector of numerics, the Z-statistics of differentail activity (DA) values for the \code{sig_gs_list}.
It is highly suggested to assign name to the vector, otherwise will use name of \code{sig_gs_list}.}

\item{use_gs2gene}{list, contains elements of gene sets. Element name is gene set name, each element contains a vector of genes belong to that gene set.
This list can be created by calling one element from \code{all_gs2gene}, or merge several gene sets into one by using \code{merge_gs}.}

\item{top_gs_number}{integer, the number of top significant gene sets to be displayed. Default is 30.}

\item{target_col}{character, name of the color palette used for display marker line in the panel.
Users can choose between "black" and "RdBu". If "black", the marker line in the panel is black. If "RdBu", the marker line in the panel is Red to Blue.
The color shade of the marker line is decided by each gene's significance of differentiation. High in red, low in blue.
Default is "RdBu".}

\item{left_annotation}{character, annotation on the left of profile curve, indicating high in control group or target group. Default is "".}

\item{right_annotation}{character, annotation on the right of profile curve, indicating high in the opposite group of \code{left_annotation}. Default is "".}

\item{main}{character, an overall title for the plot. Default is "".}

\item{profile_sig_thre}{numeric, threshold value for target genes. This parameter works only when \code{target_col_type} is set as "DE" and \code{target_col} is set as "RdBu".
Non-significant target genes will be colored grey. Default is 0.}

\item{Z_sig_thre}{numeric, threshold value of Z-statistics from \code{driver_DA_Z} and \code{driver_DE_Z}. Significant values will have background color. Default is 1.64.}

\item{pdf_file}{character, the file path to save figure as PDF file. If NULL, no PDF file will be saved. Default is NULL.}
}
\value{
Return a logical value. If TRUE, the plot has been created successfully.
}
\description{
\code{draw.GSEA.NetBID.GS} creates a GSEA plot for gene sets with more NetBID analysis information.
Such as, number of genes in each gene set, marking the rank of annotated genes in the differential expression profile and differential activity (DA) values.
}
\examples{
\dontrun{
db.preload(use_level='transcript',use_spe='human',update=FALSE)

## get all_gs2gene

analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo1','driver/DATA/',package = "NetBID2")
NetBID.loadRData(analysis.par=analysis.par,step='ms-tab')

ms_tab <- analysis.par$final_ms_tab
comp <- 'G4.Vs.others'
DE <- analysis.par$DE[[comp]]
analysis.par$out.dir.PLOT <- getwd()

## directory for saving the pdf files
exp_mat_gene <- Biobase::exprs(analysis.par$cal.eset)

## calculate activity for all genesets
use_gs2gene <- merge_gs(all_gs2gene=all_gs2gene,
                       use_gs=c('H','CP:BIOCARTA','CP:REACTOME','CP:KEGG','C5'))
ac_gs <- cal.Activity.GS(use_gs2gene = use_gs2gene,cal_mat = exp_mat_gene)

## get DA for the gene set
phe_info <- Biobase::pData(analysis.par$cal.eset)
G0  <- rownames(phe_info)[which(phe_info$`subgroup`!='G4')]
# get sample list for G0
G1  <- rownames(phe_info)[which(phe_info$`subgroup`=='G4')]
# get sample list for G1
DA_gs <- getDE.limma.2G(eset=generate.eset(ac_gs),G1=G1,G0=G0,
                        G1_name='G4',G0_name='others')
## or use: DA_gs <- getDE.BID.2G(eset=generate.eset(ac_gs),G1=G1,G0=G0,
                        G1_name='G4',G0_name='others')
## draw vocalno plot for top sig-GS
sig_gs <- draw.volcanoPlot(dat=DA_gs,
                           label_col='ID',
                           logFC_col='logFC',
                           Pv_col='P.Value',
                           logFC_thre=0.25,
                           Pv_thre=1e-4,
                           main='Volcano Plot for gene sets',
                           show_label=TRUE,
                           label_type = 'distribute',
                           label_cex = 0.5,
                           pdf_file=sprintf('\%s/vocalno_GS_DA.pdf',
                           analysis.par$out.dir.PLOT))
## GSEA plot for the significant gene sets
draw.GSEA.NetBID.GS(DE=DE,name_col='ID',
                    profile_col='t',profile_trend='pos2neg',
                    sig_gs_list = sig_gs$ID,
                    gs_DA_Z=DA_gs[sig_gs$ID,'Z-statistics'],
                    use_gs2gene = use_gs2gene,
                    top_gs_number=5,target_col='RdBu',
                    left_annotation = 'test_left',
                    right_annotation = 'test_right',
                    main='test',Z_sig_thre=1.64,profile_sig_thre = 0,
                    pdf_file=sprintf('\%s/NetBID_GSEA_GS_demo1.pdf',
                    analysis.par$out.dir.PLOT))
}
}
