% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{draw.MICA}
\alias{draw.MICA}
\title{Draw the cluster plot by MICA (cluster algorithm).}
\usage{
draw.MICA(outdir = NULL, prjname = NULL, all_k = NULL,
  obs_label = NULL, legend_pos = "topleft", legend_cex = 0.8,
  point_cex = 1, plot_type = "2D.ellipse", choose_k_strategy = "ARI",
  visualization_type = "tsne", return_type = "optimal")
}
\arguments{
\item{outdir}{character, the output directory for running MICA.}

\item{prjname}{charater, the project name set for running MICA.}

\item{all_k}{a vector of integers, the pre-defined K to evaluate.
If NULL, will use all possible K. Default is NULL.}

\item{obs_label}{a vector of characters, a vector for sample categories with names representing the sample name.}

\item{legend_pos}{character, position for the legend displayed on the plot. Default is 'topleft'.}

\item{legend_cex}{numeric, cex for the legend displayed on the plot. Default is 0.8.}

\item{point_cex}{numeric, cex for the point in the plot. Default is 1.}

\item{plot_type}{character, the type for the plot, choose from '2D' or '2D.ellipse' or '3D'. Default is '2D.ellipse'.}

\item{choose_k_strategy}{character, the strategy to choose the best K,
choose from 'ARI (adjusted rand index)', 'NMI (normalized mutual information)', 'Jaccard'. Default is 'ARI'.}

\item{visualization_type}{character,choose from 'tsne', 'umap' or 'mds'. Default is 'tsne'.}

\item{return_type}{character, the strategy to return the results, choose from 'optimal' and 'all'.
If choose 'all', cluster results from all k in \code{all_k} will be returned.
Default is 'optimal'.}
}
\value{
a vector of predicted label (if \code{return_type} is 'optimal') and a list of all possible K (if \code{return_type} is 'all')
}
\description{
\code{draw.MICA} is a visualization function to draw the cluster results for the MICA output.
}
\details{
This function mainly aims for the visulization of sample clusters.
The input is the MICA project information.
User need to input the real observation label for samples and this function will choose the best K by compared with predicted label and the observed label.
The output figure contains two sub-figures, left is labelled by the real observartion label and right is labelled by the predicted label
with comparison score (choose from ARI, NMI, Jaccard) shown above.
Not suggested when sample size is small.
}
