% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{draw.NetBID}
\alias{draw.NetBID}
\title{Plot for basic statistical display for NetBID.}
\usage{
draw.NetBID(DA_list = NULL, DE_list = NULL, main_id = NULL,
  top_number = 30, DA_display_col = "P.Value",
  DE_display_col = "logFC", z_col = "Z-statistics", digit_num = 2,
  row_cex = 1, column_cex = 1, text_cex = 1, col_srt = 60,
  pdf_file = NULL)
}
\arguments{
\item{DA_list}{list, a list of DA results.}

\item{DE_list}{list, a list of DE results.}

\item{main_id}{character, the main id for display in the figure, must be one of the name in DA_list
If NULL, will use the first name. Default is NULL.}

\item{top_number}{number for the top significant genes/drivers in the DA_list (with main_id) results to be displayed on the plot.
Default is 30.}

\item{DA_display_col}{character, column names used to display in the DA part. Default is 'P.Value'.}

\item{DE_display_col}{character, column names used to display in the DE part. Default is 'logFC'.}

\item{z_col}{character, column names for Z statistics used for background color bar. Default is 'Z-statistics'.}

\item{digit_num}{integer, number of digits to display on the plot. Default is 2.}

\item{row_cex}{numeric, \code{cex} for the row labels displayed on the plot. Default is 1}

\item{column_cex}{numeric, \code{cex} for the col labels displayed on the plot. Default is 1}

\item{text_cex}{numeric, \code{cex} for the text displayed on the plot. Default is 1}

\item{col_srt}{numeric, srt for the column labels. Default is 60.}

\item{pdf_file}{character, file path for the pdf file to save the figure into pdf format.If NULL, will not generate pdf file. Default is NULL.}
}
\value{
logical value indicating whether the plot has been successfully generated
}
\description{
\code{draw.NetBID} draw the basic statistical plot for NetBID results.
}
\examples{
analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo1','driver/DATA/',package = "NetBID2")
NetBID.loadRData(analysis.par=analysis.par,step='ms-tab')
draw.NetBID(DA_list=analysis.par$DA,DE_list=analysis.par$DE)
}
