% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{draw.network.QC}
\alias{draw.network.QC}
\title{QC Tables and Plots for Network Object}
\usage{
draw.network.QC(
  igraph_obj,
  outdir = NULL,
  prefix = "",
  directed = TRUE,
  weighted = FALSE,
  generate_html = TRUE,
  html_info_limit = TRUE
)
}
\arguments{
\item{igraph_obj}{igraph object, created by \code{get.SJAracne.network}.}

\item{outdir}{character, the output directory for saving QC tables and plots.}

\item{prefix}{character, the prefix of output QC figures names. Default is "".}

\item{directed}{logical, if TRUE, this network will be treated as a directed network. Default is TRUE.}

\item{weighted}{logical, if TRUE, this network will be treated as a weighted network. Default is FALSE.}

\item{generate_html}{logical, if TRUE, a html file will be created by R Markdown.
If FALSE, plots will be save as separated PDFs.
Default is TRUE.}

\item{html_info_limit}{logical, if TRUE, the statistics for network QC html will be limited. Default is TRUE.}
}
\value{
Return a logical value. If TRUE, success in creating QC tables and plots.
}
\description{
\code{draw.network.QC} creates tables and plots for showing some basic statistics,
driver statistics and scale-free checking of the target network.
}
\examples{
\dontrun{
if(exists('analysis.par')==TRUE) rm(analysis.par)
network.dir <- sprintf('\%s/demo1/network/',system.file(package = "NetBID2")) # use demo
network.project.name <- 'project_2019-02-14' # demo project name
project_main_dir <- 'test/'
project_name <- 'test_driver'
analysis.par  <- NetBID.analysis.dir.create(project_main_dir=project_main_dir,
                                            project_name=project_name,
                                            network_dir=network.dir,
                                            network_project_name=network.project.name)
analysis.par$tf.network  <- get.SJAracne.network(network_file=analysis.par$tf.network.file)
analysis.par$out.dir.PLOT <- getwd() ## directory for saving the pdf files
draw.network.QC(analysis.par$tf.network$igraph_obj,
                outdir=analysis.par$out.dir.QC,prefix='TF_net_')
}
}
