% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{draw.volcanoPlot}
\alias{draw.volcanoPlot}
\title{Volcano plot for top DE (differentiated expressed) genes or DA (differentiated activity) drivers}
\usage{
draw.volcanoPlot(dat = NULL, label_col = NULL, logFC_col = NULL,
  Pv_col = NULL, logFC_thre = 1.5, Pv_thre = 0.01,
  xlab = "log2 Fold Change", ylab = "P-value", show_label = FALSE,
  label_cex = 0.5, legend_cex = 0.7, label_type = "distribute",
  main = "", pdf_file = NULL)
}
\arguments{
\item{dat}{data.frame, prefer the formatted master table generated by \code{generate.masterTable}, if not, must contain the columns for the following required parameters.}

\item{label_col}{character, the name of the column in \code{dat}, which contains the gene/driver label for display}

\item{logFC_col}{character, the name of the column in \code{dat}, which contains the logFC value}

\item{Pv_col}{character, the name of the column in \code{dat}, which contains the P-value}

\item{logFC_thre}{numeric, cutoff value for the logFC. Genes or drivers with absolute logFC value higher than the cutoff are remained.Default is 1.5.}

\item{Pv_thre}{numeric, cutoff value for the p-values. Genes or drivers with lower p-values are remained.Default is 0.01.}

\item{xlab}{character, title for the X axis}

\item{ylab}{character, title for the Y axis}

\item{show_label}{logical, whether or not to display the genes or drivers passed the cutoff on the plot. Default is FALSE}

\item{label_cex}{numeric, \code{cex} for the label displayed on the plot. Default is 0.5}

\item{legend_cex}{numeric, \code{cex} for the legend displayed on the plot. Default is 0.7}

\item{label_type}{character, the strategy for label display on the plot, by choosing 'origin' or 'distribute'. Default is 'distribute'.}

\item{main}{character, \code{main} for the title on the plot.}

\item{pdf_file}{character, file path for the pdf file to save the figure into pdf format.If NULL, will not generate pdf file. Default is NULL.}
}
\value{
data.frame containing the significant genes or drivers with the following components:

\item{label_col}{}
\item{logFC_col}{}
\item{Pv_col}{}
}
\description{
\code{draw.volcanoPlot} draw the volcano plot for top DE genes or DA drivers, could display the name of the top items in figures and will retrun the list of top items.
}
\details{
This plot function input the master table and draw the volcano plot by setting significant threshold for logFC and P-value.
}
\examples{
analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo1','driver/DATA/',package = "NetBID2")
NetBID2.loadRData(analysis.par=analysis.par,step='ms-tab')
ms_tab <- analysis.par$final_ms_tab
sig_driver1 <- draw.volcanoPlot(dat=ms_tab,label_col='gene_label',
                               logFC_col='logFC.metastasis2primary_DA',
                               Pv_col='P.Value.metastasis2primary_DA',
                               logFC_thre=0.01,Pv_thre=0.1,
                               main='Volcano Plot for metastasis2primary_DA',show_label=TRUE,
                               label_cex = 1)

\dontrun{
analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo1','driver/DATA/',package = "NetBID2")
NetBID2.loadRData(analysis.par=analysis.par,step='ms-tab')
ms_tab <- analysis.par$final_ms_tab
analysis.par$out.dir.PLOT <- getwd() ## directory for saving the pdf files
sig_driver1 <- draw.volcanoPlot(dat=ms_tab,label_col='gene_label',
                               logFC_col='logFC.metastasis2primary_DA',
                               Pv_col='P.Value.metastasis2primary_DA',
                               logFC_thre=0.01,Pv_thre=0.1,
                               main='Volcano Plot for metastasis2primary_DA',
                               show_label=TRUE,
                               pdf_file=sprintf('\%s/vocalno_showlabel_distribute.pdf',
                               analysis.par$out.dir.PLOT),
                               label_cex = 1)
sig_driver2 <- draw.volcanoPlot(dat=ms_tab,label_col='gene_label',
                               logFC_col='logFC.metastasis2primary_DA',
                               Pv_col='P.Value.metastasis2primary_DA',
                               logFC_thre=0.02,Pv_thre=0.1,
                               main='Volcano Plot for metastasis2primary_DA',
                               show_label=TRUE,
                               label_type = 'origin',label_cex = 0.5,
                               pdf_file=sprintf('\%s/vocalno_showlabel_origin.pdf',
                               analysis.par$out.dir.PLOT))
sig_driver3 <- draw.volcanoPlot(dat=ms_tab,label_col='gene_label',
                               logFC_col='logFC.metastasis2primary_DA',
                               Pv_col='P.Value.metastasis2primary_DA',
                               logFC_thre=0.01,Pv_thre=0.1,
                               main='Volcano Plot for metastasis2primary_DA',
                               show_label=FALSE,
                               label_type = 'origin',label_cex = 0.5,
                               pdf_file=sprintf('\%s/vocalno_nolabel_DA.pdf',
                               analysis.par$out.dir.PLOT))
sig_gene <- draw.volcanoPlot(dat=ms_tab,label_col='geneSymbol',
                               logFC_col='logFC.metastasis2primary_DE',
                               Pv_col='P.Value.metastasis2primary_DE',
                               logFC_thre=0.5,Pv_thre=0.1,
                               main='Volcano Plot for metastasis2primary_DE',
                               show_label=FALSE,
                               pdf_file=sprintf('\%s/vocalno_nolabel_DE.pdf',
                               analysis.par$out.dir.PLOT))
}
}
