% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{funcEnrich.Fisher}
\alias{funcEnrich.Fisher}
\title{Gene set enrichment analysis by Fisher's Exact Test.}
\usage{
funcEnrich.Fisher(input_list = NULL, bg_list = NULL, use_gs = NULL,
  gs2gene = NULL, min_gs_size = 5, max_gs_size = 500,
  Pv_adj = "fdr", Pv_thre = 0.1)
}
\arguments{
\item{input_list}{a vector of characters, the list of genes for analysis. Only accept gene symbols, and gene ID conversion could be done by preparing a transfer table
by using \code{get_IDtransfer} and using \code{get_name_transfertab} to transfer the gene IDs.}

\item{bg_list}{a vector of characters, the background list of genes for analysis. Only accept gene symbols.
Default is NULL, will use all genes in the gs2gene as the background list.}

\item{use_gs}{a vector of characters, the name for gene set category used for anlaysis.
If gs2gene is set to NULL, use_gs must be the subset of \code{names(all_gs2gene)}.
Could check \code{all_gs2gene_info} for the cateogory description.
If set to 'all', all gene sets in gs2gene will be used.
Default is \code{c('H','CP:BIOCARTA','CP:REACTOME','CP:KEGG')} if gs2gene is set to NULL (use all_gs2gene).
If user input own gs2gene list, use_gs will be set to 'all' as default.}

\item{gs2gene}{a list for geneset to genes, the name for the list is the gene set name and the content in each list is the vector for genes belong to that gene set.
If NULL, will use all_gs2gene loaded by using \code{gs.preload}. Default is NULL.}

\item{min_gs_size}{numeric, minimum gene set size for analysis, default is 5.}

\item{max_gs_size}{numeric, maximum gene set size for analysis, default is 500.}

\item{Pv_adj}{character, p-value adjustment method, could check \code{p.adjust.methods} for the available options. Default is 'fdr'.}

\item{Pv_thre}{numeric, cutoff value for the adjusted p-values for significance. Default is 0.1.}
}
\value{
The function will return a list of gene sets with significant statistics, detailed as follows,

\item{#Name}{Name for the enriched gene set}
\item{Total_item}{Number of background size}
\item{Num_item}{Number of genes in the gene set (filtered by the background list)}
\item{Num_list}{Number of input genes for testing (filtered by the background list)}
\item{Num_list_item}{Number input genes annotated by the gene set (filtered by the background list)}
\item{Ori_P}{Original P-value from Fisher's Exact Test}
\item{Adj_p}{Adjusted P-value}
\item{Odds_Ratio}{Odds ratio by the 2*2 matrix used for Fisher's Exact Test}
\item{Intersected_items}{List of the intersected genes, collapsed by ';', the number is equal to Num_list_item}
}
\description{
\code{funcEnrich.Fisher} will perform gene set enrichment analysis by Fisher's Exact Test.
}
\details{
This is a function to find significant enriched gene sets for input gene list. Users could prepare gs2gene or use all_gs2gene preloaded by using \code{gs.preload}.
Background gene list is accepeted.
}
\examples{
analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo1','driver/DATA/',package = "NetBID2")
NetBID2.loadRData(analysis.par=analysis.par,step='ms-tab')
ms_tab <- analysis.par$final_ms_tab
sig_driver <- draw.volcanoPlot(dat=ms_tab,label_col='gene_label',
                               logFC_col='logFC.metastasis2primary_DA',
                               Pv_col='P.Value.metastasis2primary_DA',
                               logFC_thre=0.01,Pv_thre=0.1,
                               main='Volcano Plot for metastasis2primary_DA',
                               show_label=FALSE,
                               label_type = 'origin',label_cex = 0.5)
main.dir <- system.file(package = "NetBID2")
gs.preload(use_spe='Homo sapiens',update=FALSE)
res1 <- funcEnrich.Fisher(input_list=ms_tab[rownames(sig_driver),'geneSymbol'],
                               bg_list=ms_tab[,'geneSymbol'],
                               use_gs=c('H','C5'),
                               Pv_thre=0.1,Pv_adj = 'none')
\dontrun{
}
}
