% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{generate.eset}
\alias{generate.eset}
\title{Generate expression set.}
\usage{
generate.eset(exp_mat = NULL, phenotype_info = NULL,
  feature_info = NULL, annotation_info = "")
}
\description{
Generate expression set.
}
