% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{get.SJAracne.network}
\alias{get.SJAracne.network}
\title{get network information by input network file from SJAracne}
\usage{
get.SJAracne.network(network_file = NULL)
}
\description{
get network information by input network file from SJAracne
}
