% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{get.TF_SIG.list}
\alias{get.TF_SIG.list}
\title{Get transcription factor (TF) and signaling factor (SIG) list}
\usage{
get.TF_SIG.list(use_genes, use_gene_type = "ensembl_gene_id",
  dataset = db_info[1])
}
\description{
Get transcription factor (TF) and signaling factor (SIG) list
}
