% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{get_IDtransfer2symbol2type}
\alias{get_IDtransfer2symbol2type}
\title{Gene ID conversion related functions.}
\usage{
get_IDtransfer2symbol2type(from_type = NULL, use_genes = NULL,
  dataset = NULL, use_level = "gene")
}
\arguments{
\item{from_type}{character, attribute name from the biomaRt package,
such as 'ensembl_gene_id', 'ensembl_gene_id_version', 'ensembl_transcript_id', 'ensembl_transcript_id_version', 'refseq_mrna'.
Full list could be obtained by
\code{listAttributes(mart)$name}, where \code{mart} is the output of \code{useMart} function.
The type must be the gene type for the input \code{use_genes}.}

\item{use_genes}{a vector of characters, gene list used for ID conversion.
If NULL, will output all possible genes in transfer table. Default is NULL.}

\item{dataset}{character, name for the dataset used for ID conversion, such as 'hsapiens_gene_ensembl'.
If NULL, will use \code{db_info[1]} if run \code{db.preload} brefore. Default is NULL.}

\item{use_level}{character, either 'transcript' or 'gene', default is 'gene'}
}
\value{
\code{get_IDtransfer2symbol2type} will return a data.frame for the transfer table with gene/transcript biotype.
}
\description{
\code{get_IDtransfer2symbol2type} will generate the transfer table for the original ID to the gene symbol and gene biotype (at gene level)
or transcript symbol and transcript biotype (at transcript level).
}
\examples{
use_genes <- c("ENST00000210187","ENST00000216083","ENST00000216127",
             "ENST00000216416","ENST00000217233","ENST00000221418")
transfer_tab <- get_IDtransfer(from_type = 'ensembl_transcript_id',
                               to_type='external_gene_name',use_genes=use_genes,
                               dataset='hsapiens_gene_ensembl')
                               ## get transfer table !!!
res1 <- get_name_transfertab(use_genes,transfer_tab=transfer_tab)
transfer_tab_withtype <- get_IDtransfer2symbol2type(from_type = 'ensembl_transcript_id',
                                                   use_genes=use_genes,
                                                   dataset='hsapiens_gene_ensembl',
                                                   use_level='transcript')
                                                   ## get transfer table !!!
\dontrun{
}
}
