% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{get_obs_label}
\alias{get_obs_label}
\title{Generate a vector for sample category.}
\usage{
get_obs_label(phe_info, i)
}
\arguments{
\item{phe_info}{data.frame, phenotype data frame for the samples with sample names in rownames, e.g from \code{pData(eset)}.}

\item{i}{numeric or character, the column index or column name for extraction to get the sample category vector.}
}
\value{
Will return a vector for sample categories with names to the vector representing the sample name.
}
\description{
\code{get_obs_label} will generate a vector for sample categories with names to the vector representing the sample name.
}
\details{
This is a simple function to generate sample category vector from a data.frame.
Mainly used for input preparation in the visualization plots.
}
\examples{
analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo1','driver/DATA/',package = "NetBID2")
NetBID2.loadRData(analysis.par=analysis.par,step='ms-tab')
phe_info <- pData(analysis.par$cal.eset)
use_obs_class <- get_obs_label(phe_info = phe_info,'tumor type')
print(use_obs_class)
\dontrun{
}
}
