% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{gs.preload}
\alias{gs.preload}
\title{Load MSigDB for NetBID2 into R workspace.}
\usage{
gs.preload(use_spe = "Homo sapiens", update = FALSE, main.dir = NULL,
  db.dir = sprintf("\%s/db/", main.dir))
}
\arguments{
\item{use_spe}{character, input the species name (e.g 'Homo sapiens', 'Mus musculus'). Full list of available species name could be found by \code{msigdbr_show_species()}.
Default is 'Homo sapiens'}

\item{update}{logical,whether to update if previous RData has been generated, default FALSE}

\item{main.dir}{character, main file path for NetBID2, if NULL, will set to system.file(package = "NetBID2"). Default is NULL.}

\item{db.dir}{character, file path for saving the RData, default is \code{db} directory under \code{main.dir} when setting for \code{main.dir}.}
}
\value{
Reture TRUE if success and FALSE if not. Will load two variables into R workspace, all_gs2gene and all_gs2gene_info
}
\description{
\code{gs.preload} will load two variables into R workspace, the list for gene set to genes (all_gs2gene)
 and a dataframe (all_gs2gene_info) for detailed description for gene sets.
}
\details{
This is a pre-processing function for NetBID2 advanced analysis, user only need to input the species name (e.g Homo sapiens, Mus musculus).
The function could automatically download information from MSigDB by the functions in \code{msigdbr} and save into RData under the db/ directory
with specified species name.
}
\examples{
gs.preload(use_spe='Homo sapiens',update=FALSE)
gs.preload(use_spe='Mus musculus',update=FALSE)
print(all_gs2gene_info)
# contain the information for all gene set category, category info, category size,
## sub category,sub category info, sub category size
print(names(all_gs2gene)) # the first level of the list is the category and sub-category IDs
print(str(all_gs2gene$`CP:KEGG`))

\dontrun{
gs.preload(use_spe='Homo sapiens',update=TRUE)
}
}
