% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{load.exp.GEO}
\alias{load.exp.GEO}
\title{Load gene expression set from GEO}
\usage{
load.exp.GEO(out.dir = network.par$out.dir.DATA, GSE = NULL,
  GPL = NULL, getGPL = TRUE, update = TRUE)
}
\arguments{
\item{out.dir}{character}

\item{GSE}{character}

\item{GPL}{character}

\item{getGPL}{logical}

\item{update}{logical}
}
\description{
\code{load.exp.GEO} is a function to get GEO dataset, save into eSet and save to RData. If the RData exists, user could choose to directly load from it.
}
