% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{load.exp.GEO}
\alias{load.exp.GEO}
\title{Download Gene Expression Series From GEO Database with Platform Specified}
\usage{
load.exp.GEO(out.dir = NULL, GSE = NULL, GPL = NULL, getGPL = TRUE,
  update = FALSE)
}
\arguments{
\item{out.dir}{character, the file path used to save the GSE RData. If the data already exsits, it will be loaded from this path.}

\item{GSE}{character, the GEO Series Accession ID.}

\item{GPL}{character, the GEO Platform Accession ID.}

\item{getGPL}{logical, if TRUE, the corresponding GPL file will be downloaded. Default is TRUE.}

\item{update}{logical, if TRUE, the previous stored Gene ExpressionSet RData will be updated. Default is FALSE}
}
\value{
Return an ExpressionSet class object.
}
\description{
\code{load.exp.GEO} downloads user assigned Gene Expression Series (GSE file) along with its Platform from GEO dataset.
It returns an ExpressionSet class object and saves it as RData. If the GSE RData already exists, it will be loaded directly.
It also allows users to update the Gene Expression Series RData saved before.
}
\examples{

\dontrun{
# Download the GSE116028 which performed on GPL6480 platform from GEO and save it to the current directory
# Assign this ExpressionSet object to net_eset
net_eset <- load.exp.GEO(out.dir='./',
                         GSE='GSE116028',
                         GPL='GPL6480',
                         getGPL=TRUE,
                         update=FALSE)
}
}
