% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{load.exp.RNASeq.demo}
\alias{load.exp.RNASeq.demo}
\title{Load gene expression set from RNASeq (demo version).}
\usage{
load.exp.RNASeq.demo(files, type = "salmon", tx2gene = NULL,
  use_phenotype_info = NULL, use_sample_col = NULL,
  use_design_col = NULL, return_type = "eset", merge_level = "gene")
}
\description{
Load gene expression set from RNASeq (demo version).
}
