% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{load.exp.RNASeq.demoSalmon}
\alias{load.exp.RNASeq.demoSalmon}
\title{Load gene expression set from Salmon output (demo version).}
\usage{
load.exp.RNASeq.demoSalmon(salmon_dir = "", tx2gene = NULL,
  use_phenotype_info = NULL, use_sample_col = NULL,
  use_design_col = NULL, return_type = "tpm", merge_level = "gene")
}
\arguments{
\item{salmon_dir}{character, the main file directory to save the salmon results.}

\item{tx2gene}{data.frame or NULL, this parameter will be passed to \code{tximport}, check for detail.
If NULL, will read the transcript name in one of the salmon output
(this only works when using e.g gencode.v29.transcripts.fa from GENCODE as reference).}

\item{use_phenotype_info}{data.frame, phenotype information dataframe, must contain the columns \code{use_sample_col} and \code{use_design_col}.}

\item{use_sample_col}{character, the column name to indicate which column in \code{use_phenotype_info} should be used as the sample name.}

\item{use_design_col}{character, the column name to indicate which column in \code{use_phenotype_info} should be used as the design feature of the samples.}

\item{return_type}{character, the class of the return object, choose from 'txi','counts','tpm','eset' and 'dds'.
'txi' is the output of \code{tximport}, a simple list containing matrices: abundance, counts, length.
'counts' is the output of raw count matrix.
'tpm' is the output of raw tpm.
'dds' is the DESeqDataSet class object, the data has been processed by \code{DESeq}.
'eset' is the ExpressionSet class object,the expression data matrix has been processed by \code{DESeq} and \code{vst}.
Default is 'tpm'}

\item{merge_level}{character, choose from 'gene' and 'transcript',
if choose 'gene' and original salmon results is mapped to the transcriptome,
expression matrix will be merged to gene level.
(this only works when using e.g gencode.v29.transcripts.fa from GENCODE as reference).}
}
\description{
\code{load.exp.RNASeq.demoSalmon} is a function to read in salmon results and convert to eSet/DESeqDataSet class object.
}
\details{
This function assist to read in RNASeq results from salmon.
Due to the complicated condition (e.g reference sequence) in running salmon, this function is just a demo version that may not suit for all conditions.
}
