% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{merge_target_list}
\alias{merge_target_list}
\title{Merge Target Gene List for Two Drivers}
\usage{
merge_target_list(driver1 = NULL, driver2 = NULL, target_list = NULL)
}
\arguments{
\item{driver1}{character, the name of the first driver.}

\item{driver2}{character, the name of the second driver.}

\item{target_list}{list, the driver-to-target list object. The names of the list elements are drivers (e.g. driver1 and driver2).
Each element is a data frame, usually contains three columns. "target", target gene names; "MI", mutual information; "spearman", spearman correlation coefficient.
Users can call \code{get_net2target_list} to create this list.}
}
\value{
Return a data.frame with rows of target genes, column of "target", "MI", "spearman".
}
\description{
\code{merge_target_list} merges target gene list for two drivers together. Shared target genes with high "MI (mutual information)" statistics will be kept in the final target list.
}
\examples{
analysis.par <- list()
analysis.par$out.dir.DATA <- system.file('demo1','driver/DATA/',package = "NetBID2")
NetBID.loadRData(analysis.par=analysis.par,step='ms-tab')
ms_tab <- analysis.par$final_ms_tab
driver1 <- ms_tab[1,'originalID_label']
driver2 <- ms_tab[2,'originalID_label']
m1 <- merge_target_list(driver1=driver1,driver2=driver2,
                          target_list=analysis.par$merge.network$target_list)
}
