% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{merge_TF_SIG.AC}
\alias{merge_TF_SIG.AC}
\title{merge activity matrix for TF and SIG}
\usage{
merge_TF_SIG.AC(TF_AC = NULL, SIG_AC = NULL)
}
\arguments{
\item{TF_AC}{TF activity eset}

\item{SIG_AC}{SIG activity eset}
}
\description{
merge activity matrix for TF and SIG
}
