% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{merge_TF_SIG.network}
\alias{merge_TF_SIG.network}
\title{merge target network for TF and SIG}
\usage{
merge_TF_SIG.network(TF_network = NULL, SIG_network = NULL)
}
\arguments{
\item{TF_network}{TF network}

\item{SIG_network}{SIG network}
}
\description{
merge target network for TF and SIG
}
