% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{NetBID.analysis.dir.create}
\alias{NetBID.analysis.dir.create}
\title{Create directory for the driver analysis part. Suggested not required.}
\usage{
NetBID.analysis.dir.create(project_main_dir = NULL, prject_name = NULL,
  network_dir = NULL, network_project_name = NULL)
}
\description{
Create directory for the driver analysis part. Suggested not required.
}
