% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{NetBID.network.dir.create}
\alias{NetBID.network.dir.create}
\title{Create directory for the network generation part Suggested not required.}
\usage{
NetBID.network.dir.create(project_main_dir = NULL, prject_name = NULL)
}
\description{
Create directory for the network generation part Suggested not required.
}
