% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{out2excel}
\alias{out2excel}
\title{output master table to excel files.}
\usage{
out2excel(all_ms_tab, out.xls, mark_gene = NULL, mark_col,
  workbook_name = "ms_tab", z_column = NULL, sig_thre = 1.64)
}
\description{
output master table to excel files.
}
