% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{RNASeqCount.normalize.scale}
\alias{RNASeqCount.normalize.scale}
\title{Simple function to normalize for RNASeq Count}
\usage{
RNASeqCount.normalize.scale(d, total = NULL, pseudoCount = 1)
}
\description{
Simple function to normalize for RNASeq Count
}
