% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{SJAracne.prepare}
\alias{SJAracne.prepare}
\title{Prepare Data Files for Running SJARACNe}
\usage{
SJAracne.prepare(
  eset,
  use.samples = rownames(Biobase::pData(eset)),
  TF_list = NULL,
  SIG_list = NULL,
  SJAR.main_dir = "",
  SJAR.project_name = "",
  IQR.thre = 0.5,
  IQR.loose_thre = 0.1,
  add_options = "",
  geneSymbol_column = NULL
)
}
\arguments{
\item{eset}{an ExpressionSet class object, which contains the expression matrix.}

\item{use.samples}{a vector of characters, the list of sample used to run SJARACNe.}

\item{TF_list}{a vector of characters, the TF list.}

\item{SIG_list}{a vector of characters, the SIG list.}

\item{SJAR.main_dir}{character, the path to save the results generated by SJARACNe.}

\item{SJAR.project_name}{character, the project name used to label the output directory.}

\item{IQR.thre}{numeric, the IQR filter threshold to filter all non-driver genes.}

\item{IQR.loose_thre}{numeric, the IQR filter threshold to filter for all driver(TF/SIG) genes.}

\item{add_options}{additional option for running SJARACNe.}

\item{geneSymbol_column}{character, the column name in fdata(eset) which contains gene symbol.
If NULL, will use the main ID in exprs(eset) to fulfill the "geneSymbol" column. Default is NULL.}
}
\description{
\code{SJAracne.prepare} prepares data files for running SJAracne. SJARACNe is a scalable software tool for gene network reverse engineering from big data.
Detailed description and how to run SJARACNe can be found in its GitHub repository.
The usage of SJARACNe may be updated and the bash file generated by this function may not fit the version in use (not suit for SJAracne 0.2.0).
Please check \url{https://github.com/jyyulab/SJARACNe/} for details.
}
\examples{
\dontrun{
network.par <- list()
network.par$out.dir.DATA <- system.file('demo1','network/DATA/',package = "NetBID2")
NetBID.loadRData(network.par=network.par,step='exp-QC')
db.preload(use_level='gene',use_spe='human',update=FALSE)
use_gene_type <- 'external_gene_name' ## this should user-defined !!!
use_genes <- rownames(Biobase::fData(network.par$net.eset))
use_list  <- get.TF_SIG.list(use_genes,use_gene_type=use_gene_type)
#select sample for analysis
phe <- Biobase::pData(network.par$net.eset)
use.samples <- rownames(phe) ## use all samples, or choose to use some samples
prj.name <- network.par$project.name # can use other names, if need to run different use samples
network.par$out.dir.SJAR <- 'test' ## set the directory
SJAracne.prepare(eset=network.par$net.eset,
                 use.samples=use.samples,
                 TF_list=use_list$tf,
                 SIG_list=use_list$sig,
                 IQR.thre = 0.5,IQR.loose_thre = 0.1,
                 SJAR.project_name=prj.name,
                 SJAR.main_dir=network.par$out.dir.SJAR)
}
}
