% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{test.targetNet.overlap}
\alias{test.targetNet.overlap}
\title{Test for Intersection of Target Genes between Two Drivers}
\usage{
test.targetNet.overlap(source1_label = NULL, source2_label = NULL,
  target1 = NULL, target2 = NULL, total_possible_target = NULL)
}
\arguments{
\item{source1_label}{character, the label of the first selected driver.}

\item{source2_label}{character, the label of the second selected driver.}

\item{target1}{a vector of characters, the list of target genes from the first driver.}

\item{target2}{a vector of characters, the list of target genes from the second driver.}

\item{total_possible_target}{numeric or a vector of characters. If input is numeric, it is the total number of possible target genes.
If input is a vector of characters, it is the background list of all possible target genes.}
}
\value{
Return statistics of the testing, including the \code{P.Value}, \code{Odds_Ratio} and \code{Intersected_Number}.
}
\description{
\code{test.targetNet.overlap} performs Fisher's exact test to see whether the target genes from two drivers are significantly intersected.
}
\examples{
source1_label <- 'test1'
target1 <- sample(paste0('G',1:1000),size=80)
source2_label <- 'test2'
target2 <- sample(paste0('G',1:1000),size=120)
test.targetNet.overlap(source1_label=source1_label,source2_label=source2_label,
               target1=target1,target2=target2,
               total_possible_target=paste0('G',1:1000))
\dontrun{
}
}
