% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{update_eset.phenotype}
\alias{update_eset.phenotype}
\title{Reassign the phenoData slot of ExpressionSet and Update phenotype information}
\usage{
update_eset.phenotype(use_eset = NULL, use_phenotype_info = NULL,
  use_sample_col = NULL, use_col = "auto")
}
\arguments{
\item{use_eset}{ExpressionSet class object.}

\item{use_phenotype_info}{data.frame, a dataframe contains phenotype information, can be obtained by calling \code{pData} function.}

\item{use_sample_col}{character, must be one of the column names in \code{use_phenotype_info}.}

\item{use_col}{character, the columns will be kept from \code{use_phenotype_info}.
'auto', only extracting 'cluster-meaningful' sample features (e.g. it is meaningless to use 'gender' as clustering feature, if all samples are female).
'GEO-auto' means it will extract the following selected columns,
"geo_accession", "title", "source_name_ch1", and columns ended with ":ch1". Default is "auto".}
}
\value{
Return an ExressionSet object with updated phenotype information.
}
\description{
\code{update_eset.phenotype} reassigns the phenoData slot of ExpressionSet based on user's demand.
It is mainly used to modify sample names and extract interested phenotype information for further sample clustering.
}
\examples{
\dontrun{
net_eset <- load.exp.GEO(out.dir='./test',
                         GSE='GSE116028',
                         GPL='GPL6480',
                         getGPL=TRUE,
                         update=FALSE)
net_eset <- update_eset.phenotype(use_eset=net_eset,
                                  use_phenotype_info=pData(net_eset),
                                  use_sample_col='geo_accession',
                                  use_col='GEO-auto')
}
}
